@@ -1,50 +1,62 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
+from __future__ import annotations
 
 import sys
 
+from pathlib import Path
+from typing import TYPE_CHECKING
+
 import pytest
 
-from cleo.testers import CommandTester
+from poetry.core.semver.version import Version
 
-from poetry.core.semver import Version
-from poetry.installation.installer import Installer
 from poetry.repositories.legacy_repository import LegacyRepository
-from poetry.utils._compat import Path
 from tests.helpers import get_dependency
 from tests.helpers import get_package
 
 
-@pytest.fixture()
-def tester(app, poetry, config, executor, env):
-    tester = CommandTester(app.find("add"))
+if TYPE_CHECKING:
+    from cleo.testers.command_tester import CommandTester
+    from pytest_mock import MockerFixture
 
-    executor._io = tester.io
+    from poetry.installation.noop_installer import NoopInstaller
+    from poetry.poetry import Poetry
+    from poetry.utils.env import MockEnv
+    from poetry.utils.env import VirtualEnv
+    from tests.helpers import PoetryTestApplication
+    from tests.helpers import TestRepository
+    from tests.types import CommandTesterFactory
+    from tests.types import FixtureDirGetter
+    from tests.types import ProjectFactory
 
-    installer = Installer(
-        tester.io,
-        env,
-        poetry.package,
-        poetry.locker,
-        poetry.pool,
-        config,
-        executor=executor,
+
+@pytest.fixture
+def poetry_with_up_to_date_lockfile(
+    project_factory: ProjectFactory, fixture_dir: FixtureDirGetter
+) -> Poetry:
+    source = fixture_dir("up_to_date_lock")
+
+    return project_factory(
+        name="foobar",
+        pyproject_content=(source / "pyproject.toml").read_text(encoding="utf-8"),
+        poetry_lock_content=(source / "poetry.lock").read_text(encoding="utf-8"),
     )
-    installer.use_executor(True)
-    tester._command.set_installer(installer)
-    tester._command.set_env(env)
-
-    return tester
 
 
 @pytest.fixture()
-def old_tester(tester):
-    tester._command.installer.use_executor(False)
+def tester(command_tester_factory: CommandTesterFactory) -> CommandTester:
+    return command_tester_factory("add")
+
+
+@pytest.fixture()
+def old_tester(tester: CommandTester) -> CommandTester:
+    tester.command.installer.use_executor(False)
 
     return tester
 
 
-def test_add_no_constraint(app, repo, tester):
+def test_add_no_constraint(
+    app: PoetryTestApplication, repo: TestRepository, tester: CommandTester
+):
     repo.add_package(get_package("cachy", "0.1.0"))
     repo.add_package(get_package("cachy", "0.2.0"))
 
@@ -63,8 +75,8 @@ Package operations: 1 install, 0 updates, 0 removals
   • Installing cachy (0.2.0)
 """
 
-    assert expected == tester.io.fetch_output()
-    assert 1 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 1
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
@@ -72,7 +84,77 @@ Package operations: 1 install, 0 updates, 0 removals
     assert content["dependencies"]["cachy"] == "^0.2.0"
 
 
-def test_add_equal_constraint(app, repo, tester):
+def test_add_replace_by_constraint(
+    app: PoetryTestApplication, repo: TestRepository, tester: CommandTester
+):
+    repo.add_package(get_package("cachy", "0.1.0"))
+    repo.add_package(get_package("cachy", "0.2.0"))
+
+    tester.execute("cachy")
+
+    expected = """\
+Using version ^0.2.0 for cachy
+
+Updating dependencies
+Resolving dependencies...
+
+Writing lock file
+
+Package operations: 1 install, 0 updates, 0 removals
+
+  • Installing cachy (0.2.0)
+"""
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 1
+
+    content = app.poetry.file.read()["tool"]["poetry"]
+
+    assert "cachy" in content["dependencies"]
+    assert content["dependencies"]["cachy"] == "^0.2.0"
+
+    tester.execute("cachy@0.1.0")
+    expected = """
+Updating dependencies
+Resolving dependencies...
+
+Writing lock file
+
+Package operations: 1 install, 0 updates, 0 removals
+
+  • Installing cachy (0.1.0)
+"""
+    assert tester.io.fetch_output() == expected
+
+    content = app.poetry.file.read()["tool"]["poetry"]
+
+    assert "cachy" in content["dependencies"]
+    assert content["dependencies"]["cachy"] == "0.1.0"
+
+
+def test_add_no_constraint_editable_error(
+    app: PoetryTestApplication, repo: TestRepository, tester: CommandTester
+):
+    content = app.poetry.file.read()["tool"]["poetry"]
+
+    repo.add_package(get_package("cachy", "0.2.0"))
+
+    tester.execute("-e cachy")
+
+    expected = """
+Failed to add packages. Only vcs/path dependencies support editable installs.\
+ cachy is neither.
+
+No changes were applied.
+"""
+    assert tester.status_code == 1
+    assert tester.io.fetch_error() == expected
+    assert tester.command.installer.executor.installations_count == 0
+    assert content == app.poetry.file.read()["tool"]["poetry"]
+
+
+def test_add_equal_constraint(
+    app: PoetryTestApplication, repo: TestRepository, tester: CommandTester
+):
     repo.add_package(get_package("cachy", "0.1.0"))
     repo.add_package(get_package("cachy", "0.2.0"))
 
@@ -90,11 +172,13 @@ Package operations: 1 install, 0 updates, 0 removals
   • Installing cachy (0.1.0)
 """
 
-    assert expected == tester.io.fetch_output()
-    assert 1 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 1
 
 
-def test_add_greater_constraint(app, repo, tester):
+def test_add_greater_constraint(
+    app: PoetryTestApplication, repo: TestRepository, tester: CommandTester
+):
     repo.add_package(get_package("cachy", "0.1.0"))
     repo.add_package(get_package("cachy", "0.2.0"))
 
@@ -112,21 +196,27 @@ Package operations: 1 install, 0 updates, 0 removals
   • Installing cachy (0.2.0)
 """
 
-    assert expected == tester.io.fetch_output()
-    assert 1 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 1
 
 
-def test_add_constraint_with_extras(app, repo, tester):
+@pytest.mark.parametrize("extra_name", ["msgpack", "MsgPack"])
+def test_add_constraint_with_extras(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    tester: CommandTester,
+    extra_name: str,
+):
     cachy1 = get_package("cachy", "0.1.0")
     cachy1.extras = {"msgpack": [get_dependency("msgpack-python")]}
     msgpack_dep = get_dependency("msgpack-python", ">=0.5 <0.6", optional=True)
-    cachy1.requires = [msgpack_dep]
+    cachy1.add_dependency(msgpack_dep)
 
     repo.add_package(get_package("cachy", "0.2.0"))
     repo.add_package(cachy1)
     repo.add_package(get_package("msgpack-python", "0.5.3"))
 
-    tester.execute("cachy[msgpack]>=0.1.0,<0.2.0")
+    tester.execute(f"cachy[{extra_name}]>=0.1.0,<0.2.0")
 
     expected = """\
 
@@ -141,14 +231,16 @@ Package operations: 2 installs, 0 updates, 0 removals
   • Installing cachy (0.1.0)
 """
 
-    assert expected == tester.io.fetch_output()
-    assert 2 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 2
 
 
-def test_add_constraint_dependencies(app, repo, tester):
+def test_add_constraint_dependencies(
+    app: PoetryTestApplication, repo: TestRepository, tester: CommandTester
+):
     cachy2 = get_package("cachy", "0.2.0")
     msgpack_dep = get_dependency("msgpack-python", ">=0.5 <0.6")
-    cachy2.requires = [msgpack_dep]
+    cachy2.add_dependency(msgpack_dep)
 
     repo.add_package(get_package("cachy", "0.1.0"))
     repo.add_package(cachy2)
@@ -169,12 +261,17 @@ Package operations: 2 installs, 0 updates, 0 removals
   • Installing cachy (0.2.0)
 """
 
-    assert expected == tester.io.fetch_output()
-    assert 2 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 2
 
 
-def test_add_git_constraint(app, repo, tester, tmp_venv):
-    tester._command.set_env(tmp_venv)
+def test_add_git_constraint(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    tester: CommandTester,
+    tmp_venv: VirtualEnv,
+):
+    tester.command.set_env(tmp_venv)
 
     repo.add_package(get_package("pendulum", "1.4.4"))
     repo.add_package(get_package("cleo", "0.6.5"))
@@ -194,8 +291,8 @@ Package operations: 2 installs, 0 updates, 0 removals
   • Installing demo (0.1.2 9cf87a2)
 """
 
-    assert expected == tester.io.fetch_output()
-    assert 2 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 2
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
@@ -205,8 +302,13 @@ Package operations: 2 installs, 0 updates, 0 removals
     }
 
 
-def test_add_git_constraint_with_poetry(app, repo, tester, tmp_venv):
-    tester._command.set_env(tmp_venv)
+def test_add_git_constraint_with_poetry(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    tester: CommandTester,
+    tmp_venv: VirtualEnv,
+):
+    tester.command.set_env(tmp_venv)
 
     repo.add_package(get_package("pendulum", "1.4.4"))
 
@@ -225,18 +327,25 @@ Package operations: 2 installs, 0 updates, 0 removals
   • Installing demo (0.1.2 9cf87a2)
 """
 
-    assert expected == tester.io.fetch_output()
-    assert 2 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 2
 
 
-def test_add_git_constraint_with_extras(app, repo, tester, tmp_venv):
-    tester._command.set_env(tmp_venv)
+@pytest.mark.parametrize("extra_name", ["foo", "FOO"])
+def test_add_git_constraint_with_extras(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    tester: CommandTester,
+    tmp_venv: VirtualEnv,
+    extra_name: str,
+):
+    tester.command.set_env(tmp_venv)
 
     repo.add_package(get_package("pendulum", "1.4.4"))
     repo.add_package(get_package("cleo", "0.6.5"))
     repo.add_package(get_package("tomlkit", "0.5.5"))
 
-    tester.execute("git+https://github.com/demo/demo.git[foo,bar]")
+    tester.execute(f"git+https://github.com/demo/demo.git[{extra_name},bar]")
 
     expected = """\
 
@@ -253,25 +362,80 @@ Package operations: 4 installs, 0 updates, 0 removals
   • Installing demo (0.1.2 9cf87a2)
 """
 
-    assert expected.strip() == tester.io.fetch_output().strip()
-    assert 4 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output().strip() == expected.strip()
+    assert tester.command.installer.executor.installations_count == 4
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
     assert "demo" in content["dependencies"]
     assert content["dependencies"]["demo"] == {
         "git": "https://github.com/demo/demo.git",
-        "extras": ["foo", "bar"],
+        "extras": [extra_name, "bar"],
     }
 
 
-def test_add_git_ssh_constraint(app, repo, tester, tmp_venv):
-    tester._command.set_env(tmp_venv)
+@pytest.mark.parametrize(
+    "url, rev",
+    [
+        ("git+https://github.com/demo/subdirectories.git#subdirectory=two", None),
+        (
+            "git+https://github.com/demo/subdirectories.git@master#subdirectory=two",
+            "master",
+        ),
+    ],
+)
+def test_add_git_constraint_with_subdirectory(
+    url: str,
+    rev: str | None,
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    tester: CommandTester,
+    env: MockEnv,
+):
+    tester.execute(url)
+
+    expected = """\
+Updating dependencies
+Resolving dependencies...
+
+Writing lock file
+
+Package operations: 1 install, 0 updates, 0 removals
+
+  • Installing two (2.0.0 9cf87a2)
+"""
+    assert tester.io.fetch_output().strip() == expected.strip()
+    assert tester.command.installer.executor.installations_count == 1
+
+    content = app.poetry.file.read()["tool"]["poetry"]
+
+    constraint = {
+        "git": "https://github.com/demo/subdirectories.git",
+        "subdirectory": "two",
+    }
+
+    if rev:
+        constraint["rev"] = rev
+
+    assert "two" in content["dependencies"]
+    assert content["dependencies"]["two"] == constraint
+
+
+@pytest.mark.parametrize("editable", [False, True])
+def test_add_git_ssh_constraint(
+    editable: bool,
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    tester: CommandTester,
+    tmp_venv: VirtualEnv,
+):
+    tester.command.set_env(tmp_venv)
 
     repo.add_package(get_package("pendulum", "1.4.4"))
     repo.add_package(get_package("cleo", "0.6.5"))
 
-    tester.execute("git+ssh://git@github.com/demo/demo.git@develop")
+    url = "git+ssh://git@github.com/demo/demo.git@develop"
+    tester.execute(f"{url}" if not editable else f"-e {url}")
 
     expected = """\
 
@@ -286,29 +450,41 @@ Package operations: 2 installs, 0 updates, 0 removals
   • Installing demo (0.1.2 9cf87a2)
 """
 
-    assert expected == tester.io.fetch_output()
-    assert 2 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 2
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
     assert "demo" in content["dependencies"]
-    assert content["dependencies"]["demo"] == {
+
+    expected = {
         "git": "ssh://git@github.com/demo/demo.git",
         "rev": "develop",
     }
+    if editable:
+        expected["develop"] = True
+
+    assert content["dependencies"]["demo"] == expected
 
 
-def test_add_directory_constraint(app, repo, tester, mocker):
-    p = mocker.patch("poetry.utils._compat.Path.cwd")
-    p.return_value = Path(__file__).parent
-
+@pytest.mark.parametrize(
+    "required_fixtures",
+    [["git/github.com/demo/demo"]],
+)
+@pytest.mark.parametrize("editable", [False, True])
+def test_add_directory_constraint(
+    editable: bool,
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    tester: CommandTester,
+):
     repo.add_package(get_package("pendulum", "1.4.4"))
     repo.add_package(get_package("cleo", "0.6.5"))
 
     path = "../git/github.com/demo/demo"
-    tester.execute("{}".format(path))
+    tester.execute(f"{path}" if not editable else f"-e {path}")
 
-    expected = """\
+    expected = f"""\
 
 Updating dependencies
 Resolving dependencies...
@@ -318,30 +494,38 @@ Writing lock file
 Package operations: 2 installs, 0 updates, 0 removals
 
   • Installing pendulum (1.4.4)
-  • Installing demo (0.1.2 {})
-""".format(
-        app.poetry.file.parent.joinpath(path).resolve().as_posix()
-    )
+  • Installing demo (0.1.2 {app.poetry.file.parent.joinpath(path).resolve().as_posix()})
+"""
 
-    assert expected == tester.io.fetch_output()
-    assert 2 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 2
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
     assert "demo" in content["dependencies"]
-    assert content["dependencies"]["demo"] == {"path": "../git/github.com/demo/demo"}
+
+    expected = {"path": path}
+    if editable:
+        expected["develop"] = True
+
+    assert content["dependencies"]["demo"] == expected
 
 
-def test_add_directory_with_poetry(app, repo, tester, mocker):
-    p = mocker.patch("poetry.utils._compat.Path.cwd")
-    p.return_value = Path(__file__) / ".."
-
+@pytest.mark.parametrize(
+    "required_fixtures",
+    [["git/github.com/demo/pyproject-demo"]],
+)
+def test_add_directory_with_poetry(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    tester: CommandTester,
+):
     repo.add_package(get_package("pendulum", "1.4.4"))
 
     path = "../git/github.com/demo/pyproject-demo"
-    tester.execute("{}".format(path))
+    tester.execute(f"{path}")
 
-    expected = """\
+    expected = f"""\
 
 Updating dependencies
 Resolving dependencies...
@@ -351,25 +535,29 @@ Writing lock file
 Package operations: 2 installs, 0 updates, 0 removals
 
   • Installing pendulum (1.4.4)
-  • Installing demo (0.1.2 {})
-""".format(
-        app.poetry.file.parent.joinpath(path).resolve().as_posix()
-    )
+  • Installing demo (0.1.2 {app.poetry.file.parent.joinpath(path).resolve().as_posix()})
+"""
 
-    assert expected == tester.io.fetch_output()
-    assert 2 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 2
 
 
-def test_add_file_constraint_wheel(app, repo, tester, mocker, poetry):
-    p = mocker.patch("poetry.utils._compat.Path.cwd")
-    p.return_value = poetry.file.parent
-
+@pytest.mark.parametrize(
+    "required_fixtures",
+    [["distributions/demo-0.1.0-py2.py3-none-any.whl"]],
+)
+def test_add_file_constraint_wheel(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    tester: CommandTester,
+    poetry: Poetry,
+):
     repo.add_package(get_package("pendulum", "1.4.4"))
 
     path = "../distributions/demo-0.1.0-py2.py3-none-any.whl"
-    tester.execute("{}".format(path))
+    tester.execute(f"{path}")
 
-    expected = """\
+    expected = f"""\
 
 Updating dependencies
 Resolving dependencies...
@@ -379,32 +567,33 @@ Writing lock file
 Package operations: 2 installs, 0 updates, 0 removals
 
   • Installing pendulum (1.4.4)
-  • Installing demo (0.1.0 {})
-""".format(
-        app.poetry.file.parent.joinpath(path).resolve().as_posix()
-    )
+  • Installing demo (0.1.0 {app.poetry.file.parent.joinpath(path).resolve().as_posix()})
+"""
 
-    assert expected == tester.io.fetch_output()
-    assert 2 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 2
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
     assert "demo" in content["dependencies"]
-    assert content["dependencies"]["demo"] == {
-        "path": "../distributions/demo-0.1.0-py2.py3-none-any.whl"
-    }
+    assert content["dependencies"]["demo"] == {"path": path}
 
 
-def test_add_file_constraint_sdist(app, repo, tester, mocker):
-    p = mocker.patch("poetry.utils._compat.Path.cwd")
-    p.return_value = Path(__file__) / ".."
-
+@pytest.mark.parametrize(
+    "required_fixtures",
+    [["distributions/demo-0.1.0.tar.gz"]],
+)
+def test_add_file_constraint_sdist(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    tester: CommandTester,
+):
     repo.add_package(get_package("pendulum", "1.4.4"))
 
     path = "../distributions/demo-0.1.0.tar.gz"
-    tester.execute("{}".format(path))
+    tester.execute(f"{path}")
 
-    expected = """\
+    expected = f"""\
 
 Updating dependencies
 Resolving dependencies...
@@ -414,33 +603,35 @@ Writing lock file
 Package operations: 2 installs, 0 updates, 0 removals
 
   • Installing pendulum (1.4.4)
-  • Installing demo (0.1.0 {})
-""".format(
-        app.poetry.file.parent.joinpath(path).resolve().as_posix()
-    )
+  • Installing demo (0.1.0 {app.poetry.file.parent.joinpath(path).resolve().as_posix()})
+"""
 
-    assert expected == tester.io.fetch_output()
-    assert 2 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 2
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
     assert "demo" in content["dependencies"]
-    assert content["dependencies"]["demo"] == {
-        "path": "../distributions/demo-0.1.0.tar.gz"
-    }
+    assert content["dependencies"]["demo"] == {"path": path}
 
 
-def test_add_constraint_with_extras_option(app, repo, tester):
+@pytest.mark.parametrize("extra_name", ["msgpack", "MsgPack"])
+def test_add_constraint_with_extras_option(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    tester: CommandTester,
+    extra_name: str,
+):
     cachy2 = get_package("cachy", "0.2.0")
     cachy2.extras = {"msgpack": [get_dependency("msgpack-python")]}
     msgpack_dep = get_dependency("msgpack-python", ">=0.5 <0.6", optional=True)
-    cachy2.requires = [msgpack_dep]
+    cachy2.add_dependency(msgpack_dep)
 
     repo.add_package(get_package("cachy", "0.1.0"))
     repo.add_package(cachy2)
     repo.add_package(get_package("msgpack-python", "0.5.3"))
 
-    tester.execute("cachy=0.2.0 --extras msgpack")
+    tester.execute(f"cachy=0.2.0 --extras {extra_name}")
 
     expected = """\
 
@@ -455,20 +646,25 @@ Package operations: 2 installs, 0 updates, 0 removals
   • Installing cachy (0.2.0)
 """
 
-    assert expected == tester.io.fetch_output()
-    assert 2 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 2
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
     assert "cachy" in content["dependencies"]
     assert content["dependencies"]["cachy"] == {
         "version": "0.2.0",
-        "extras": ["msgpack"],
+        "extras": [extra_name],
     }
 
 
-def test_add_url_constraint_wheel(app, repo, tester, mocker):
-    p = mocker.patch("poetry.utils._compat.Path.cwd")
+def test_add_url_constraint_wheel(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    tester: CommandTester,
+    mocker: MockerFixture,
+):
+    p = mocker.patch("pathlib.Path.cwd")
     p.return_value = Path(__file__) / ".."
 
     repo.add_package(get_package("pendulum", "1.4.4"))
@@ -487,11 +683,12 @@ Writing lock file
 Package operations: 2 installs, 0 updates, 0 removals
 
   • Installing pendulum (1.4.4)
-  • Installing demo (0.1.0 https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl)
+  • Installing demo\
+ (0.1.0 https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl)
 """
 
-    assert expected == tester.io.fetch_output()
-    assert 2 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 2
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
@@ -501,13 +698,21 @@ Package operations: 2 installs, 0 updates, 0 removals
     }
 
 
-def test_add_url_constraint_wheel_with_extras(app, repo, tester, mocker):
+@pytest.mark.parametrize("extra_name", ["foo", "FOO"])
+def test_add_url_constraint_wheel_with_extras(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    tester: CommandTester,
+    extra_name: str,
+    mocker: MockerFixture,
+):
     repo.add_package(get_package("pendulum", "1.4.4"))
     repo.add_package(get_package("cleo", "0.6.5"))
     repo.add_package(get_package("tomlkit", "0.5.5"))
 
     tester.execute(
-        "https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl[foo,bar]"
+        "https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl"
+        f"[{extra_name},bar]"
     )
 
     expected = """\
@@ -522,24 +727,29 @@ Package operations: 4 installs, 0 updates, 0 removals
   • Installing cleo (0.6.5)
   • Installing pendulum (1.4.4)
   • Installing tomlkit (0.5.5)
-  • Installing demo (0.1.0 https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl)
+  • Installing demo\
+ (0.1.0 https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl)
 """
     # Order might be different, split into lines and compare the overall output.
     expected = set(expected.splitlines())
     output = set(tester.io.fetch_output().splitlines())
-    assert expected == output
-    assert 4 == tester._command.installer.executor.installations_count
+    assert output == expected
+    assert tester.command.installer.executor.installations_count == 4
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
     assert "demo" in content["dependencies"]
     assert content["dependencies"]["demo"] == {
-        "url": "https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl",
-        "extras": ["foo", "bar"],
+        "url": (
+            "https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl"
+        ),
+        "extras": [extra_name, "bar"],
     }
 
 
-def test_add_constraint_with_python(app, repo, tester):
+def test_add_constraint_with_python(
+    app: PoetryTestApplication, repo: TestRepository, tester: CommandTester
+):
     cachy2 = get_package("cachy", "0.2.0")
 
     repo.add_package(get_package("cachy", "0.1.0"))
@@ -559,8 +769,8 @@ Package operations: 1 install, 0 updates, 0 removals
   • Installing cachy (0.2.0)
 """
 
-    assert expected == tester.io.fetch_output()
-    assert 1 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 1
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
@@ -568,7 +778,12 @@ Package operations: 1 install, 0 updates, 0 removals
     assert content["dependencies"]["cachy"] == {"version": "0.2.0", "python": ">=2.7"}
 
 
-def test_add_constraint_with_platform(app, repo, tester, env):
+def test_add_constraint_with_platform(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    tester: CommandTester,
+    env: MockEnv,
+):
     platform = sys.platform
     env._platform = platform
 
@@ -577,7 +792,7 @@ def test_add_constraint_with_platform(app, repo, tester, env):
     repo.add_package(get_package("cachy", "0.1.0"))
     repo.add_package(cachy2)
 
-    tester.execute("cachy=0.2.0 --platform {} -vvv".format(platform))
+    tester.execute(f"cachy=0.2.0 --platform {platform} -vvv")
 
     expected = """\
 
@@ -591,8 +806,8 @@ Package operations: 1 install, 0 updates, 0 removals
   • Installing cachy (0.2.0)
 """
 
-    assert expected == tester.io.fetch_output()
-    assert 1 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 1
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
@@ -603,7 +818,9 @@ Package operations: 1 install, 0 updates, 0 removals
     }
 
 
-def test_add_constraint_with_source(app, poetry, tester):
+def test_add_constraint_with_source(
+    app: PoetryTestApplication, poetry: Poetry, tester: CommandTester
+):
     repo = LegacyRepository(name="my-index", url="https://my-index.fake")
     repo.add_package(get_package("cachy", "0.2.0"))
     repo._cache.store("matches").put("cachy:0.2.0", [Version.parse("0.2.0")], 5)
@@ -624,8 +841,8 @@ Package operations: 1 install, 0 updates, 0 removals
   • Installing cachy (0.2.0)
 """
 
-    assert expected == tester.io.fetch_output()
-    assert 1 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 1
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
@@ -636,14 +853,21 @@ Package operations: 1 install, 0 updates, 0 removals
     }
 
 
-def test_add_constraint_with_source_that_does_not_exist(app, tester):
+def test_add_constraint_with_source_that_does_not_exist(
+    app: PoetryTestApplication, tester: CommandTester
+):
     with pytest.raises(ValueError) as e:
         tester.execute("foo --source i-dont-exist")
 
-    assert 'Repository "i-dont-exist" does not exist.' == str(e.value)
+    assert str(e.value) == 'Repository "i-dont-exist" does not exist.'
 
 
-def test_add_constraint_not_found_with_source(app, poetry, mocker, tester):
+def test_add_constraint_not_found_with_source(
+    app: PoetryTestApplication,
+    poetry: Poetry,
+    mocker: MockerFixture,
+    tester: CommandTester,
+):
     repo = LegacyRepository(name="my-index", url="https://my-index.fake")
     mocker.patch.object(repo, "find_packages", return_value=[])
 
@@ -655,14 +879,16 @@ def test_add_constraint_not_found_with_source(app, poetry, mocker, tester):
     with pytest.raises(ValueError) as e:
         tester.execute("cachy --source my-index")
 
-    assert "Could not find a matching version of package cachy" == str(e.value)
+    assert str(e.value) == "Could not find a matching version of package cachy"
 
 
-def test_add_to_section_that_does_no_exist_yet(app, repo, tester):
+def test_add_to_section_that_does_not_exist_yet(
+    app: PoetryTestApplication, repo: TestRepository, tester: CommandTester
+):
     repo.add_package(get_package("cachy", "0.1.0"))
     repo.add_package(get_package("cachy", "0.2.0"))
 
-    tester.execute("cachy --dev")
+    tester.execute("cachy --group dev")
 
     expected = """\
 Using version ^0.2.0 for cachy
@@ -677,16 +903,66 @@ Package operations: 1 install, 0 updates, 0 removals
   • Installing cachy (0.2.0)
 """
 
-    assert expected == tester.io.fetch_output()
-    assert 1 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 1
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
-    assert "cachy" in content["dev-dependencies"]
-    assert content["dev-dependencies"]["cachy"] == "^0.2.0"
+    assert "cachy" in content["group"]["dev"]["dependencies"]
+    assert content["group"]["dev"]["dependencies"]["cachy"] == "^0.2.0"
+
+    expected = """\
+
+[tool.poetry.group.dev.dependencies]
+cachy = "^0.2.0"
+
+"""
+    string_content = content.as_string()
+    if "\r\n" in string_content:
+        # consistent line endings
+        expected = expected.replace("\n", "\r\n")
+
+    assert expected in string_content
 
 
-def test_add_should_not_select_prereleases(app, repo, tester):
+def test_add_to_dev_section_deprecated(
+    app: PoetryTestApplication, repo: TestRepository, tester: CommandTester
+):
+    repo.add_package(get_package("cachy", "0.1.0"))
+    repo.add_package(get_package("cachy", "0.2.0"))
+
+    tester.execute("cachy --dev")
+
+    warning = """\
+The --dev option is deprecated, use the `--group dev` notation instead.
+"""
+
+    expected = """\
+Using version ^0.2.0 for cachy
+
+Updating dependencies
+Resolving dependencies...
+
+Writing lock file
+
+Package operations: 1 install, 0 updates, 0 removals
+
+  • Installing cachy (0.2.0)
+"""
+
+    assert tester.io.fetch_error() == warning
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 1
+
+    content = app.poetry.file.read()["tool"]["poetry"]
+
+    assert "cachy" in content["group"]["dev"]["dependencies"]
+    assert content["group"]["dev"]["dependencies"]["cachy"] == "^0.2.0"
+
+
+def test_add_should_not_select_prereleases(
+    app: PoetryTestApplication, repo: TestRepository, tester: CommandTester
+):
     repo.add_package(get_package("pyyaml", "3.13"))
     repo.add_package(get_package("pyyaml", "4.2b2"))
 
@@ -705,8 +981,8 @@ Package operations: 1 install, 0 updates, 0 removals
   • Installing pyyaml (3.13)
 """
 
-    assert expected == tester.io.fetch_output()
-    assert 1 == tester._command.installer.executor.installations_count
+    assert tester.io.fetch_output() == expected
+    assert tester.command.installer.executor.installations_count == 1
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
@@ -714,23 +990,56 @@ Package operations: 1 install, 0 updates, 0 removals
     assert content["dependencies"]["pyyaml"] == "^3.13"
 
 
-def test_add_should_display_an_error_when_adding_existing_package_with_no_constraint(
-    app, repo, tester
+def test_add_should_skip_when_adding_existing_package_with_no_constraint(
+    app: PoetryTestApplication, repo: TestRepository, tester: CommandTester
 ):
     content = app.poetry.file.read()
     content["tool"]["poetry"]["dependencies"]["foo"] = "^1.0"
     app.poetry.file.write(content)
 
     repo.add_package(get_package("foo", "1.1.2"))
+    tester.execute("foo")
 
-    with pytest.raises(ValueError) as e:
-        tester.execute("foo")
+    expected = """\
+The following packages are already present in the pyproject.toml and will be skipped:
 
-    assert "Package foo is already present" == str(e.value)
+  • foo
+
+If you want to update it to the latest compatible version,\
+ you can use `poetry update package`.
+If you prefer to upgrade it to the latest available version,\
+ you can use `poetry add package@latest`.
+"""
+
+    assert expected in tester.io.fetch_output()
+
+
+def test_add_should_skip_when_adding_canonicalized_existing_package_with_no_constraint(
+    app: PoetryTestApplication, repo: TestRepository, tester: CommandTester
+):
+    content = app.poetry.file.read()
+    content["tool"]["poetry"]["dependencies"]["foo-bar"] = "^1.0"
+    app.poetry.file.write(content)
+
+    repo.add_package(get_package("foo-bar", "1.1.2"))
+    tester.execute("Foo_Bar")
+
+    expected = """\
+The following packages are already present in the pyproject.toml and will be skipped:
+
+  • Foo_Bar
+
+If you want to update it to the latest compatible version,\
+ you can use `poetry update package`.
+If you prefer to upgrade it to the latest available version,\
+ you can use `poetry add package@latest`.
+"""
+
+    assert expected in tester.io.fetch_output()
 
 
 def test_add_should_work_when_adding_existing_package_with_latest_constraint(
-    app, repo, tester
+    app: PoetryTestApplication, repo: TestRepository, tester: CommandTester
 ):
     content = app.poetry.file.read()
     content["tool"]["poetry"]["dependencies"]["foo"] = "^1.0"
@@ -761,14 +1070,16 @@ Package operations: 1 install, 0 updates, 0 removals
     assert content["dependencies"]["foo"] == "^1.1.2"
 
 
-def test_add_chooses_prerelease_if_only_prereleases_are_available(app, repo, tester):
+def test_add_chooses_prerelease_if_only_prereleases_are_available(
+    app: PoetryTestApplication, repo: TestRepository, tester: CommandTester
+):
     repo.add_package(get_package("foo", "1.2.3b0"))
     repo.add_package(get_package("foo", "1.2.3b1"))
 
     tester.execute("foo")
 
     expected = """\
-Using version ^1.2.3-beta.1 for foo
+Using version ^1.2.3b1 for foo
 
 Updating dependencies
 Resolving dependencies...
@@ -779,11 +1090,12 @@ Package operations: 1 install, 0 updates, 0 removals
 
   • Installing foo (1.2.3b1)
 """
-
     assert expected in tester.io.fetch_output()
 
 
-def test_add_prefers_stable_releases(app, repo, tester):
+def test_add_prefers_stable_releases(
+    app: PoetryTestApplication, repo: TestRepository, tester: CommandTester
+):
     repo.add_package(get_package("foo", "1.2.3"))
     repo.add_package(get_package("foo", "1.2.4b1"))
 
@@ -805,7 +1117,9 @@ Package operations: 1 install, 0 updates, 0 removals
     assert expected in tester.io.fetch_output()
 
 
-def test_add_with_lock(app, repo, tester):
+def test_add_with_lock(
+    app: PoetryTestApplication, repo: TestRepository, tester: CommandTester
+):
     content_hash = app.poetry.locker._get_content_hash()
     repo.add_package(get_package("cachy", "0.2.0"))
 
@@ -820,11 +1134,16 @@ Resolving dependencies...
 Writing lock file
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
     assert content_hash != app.poetry.locker.lock_data["metadata"]["content-hash"]
 
 
-def test_add_no_constraint_old_installer(app, repo, installer, old_tester):
+def test_add_no_constraint_old_installer(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
+):
     repo.add_package(get_package("cachy", "0.1.0"))
     repo.add_package(get_package("cachy", "0.2.0"))
 
@@ -843,7 +1162,7 @@ Package operations: 1 install, 0 updates, 0 removals
   - Installing cachy (0.2.0)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 1
 
@@ -853,7 +1172,12 @@ Package operations: 1 install, 0 updates, 0 removals
     assert content["dependencies"]["cachy"] == "^0.2.0"
 
 
-def test_add_equal_constraint_old_installer(app, repo, installer, old_tester):
+def test_add_equal_constraint_old_installer(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
+):
     repo.add_package(get_package("cachy", "0.1.0"))
     repo.add_package(get_package("cachy", "0.2.0"))
 
@@ -871,12 +1195,17 @@ Package operations: 1 install, 0 updates, 0 removals
   - Installing cachy (0.1.0)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 1
 
 
-def test_add_greater_constraint_old_installer(app, repo, installer, old_tester):
+def test_add_greater_constraint_old_installer(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
+):
     repo.add_package(get_package("cachy", "0.1.0"))
     repo.add_package(get_package("cachy", "0.2.0"))
 
@@ -894,22 +1223,29 @@ Package operations: 1 install, 0 updates, 0 removals
   - Installing cachy (0.2.0)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 1
 
 
-def test_add_constraint_with_extras_old_installer(app, repo, installer, old_tester):
+@pytest.mark.parametrize("extra_name", ["msgpack", "MsgPack"])
+def test_add_constraint_with_extras_old_installer(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
+    extra_name: str,
+):
     cachy1 = get_package("cachy", "0.1.0")
     cachy1.extras = {"msgpack": [get_dependency("msgpack-python")]}
     msgpack_dep = get_dependency("msgpack-python", ">=0.5 <0.6", optional=True)
-    cachy1.requires = [msgpack_dep]
+    cachy1.add_dependency(msgpack_dep)
 
     repo.add_package(get_package("cachy", "0.2.0"))
     repo.add_package(cachy1)
     repo.add_package(get_package("msgpack-python", "0.5.3"))
 
-    old_tester.execute("cachy[msgpack]>=0.1.0,<0.2.0")
+    old_tester.execute(f"cachy[{extra_name}]>=0.1.0,<0.2.0")
 
     expected = """\
 
@@ -924,15 +1260,20 @@ Package operations: 2 installs, 0 updates, 0 removals
   - Installing cachy (0.1.0)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 2
 
 
-def test_add_constraint_dependencies_old_installer(app, repo, installer, old_tester):
+def test_add_constraint_dependencies_old_installer(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
+):
     cachy2 = get_package("cachy", "0.2.0")
     msgpack_dep = get_dependency("msgpack-python", ">=0.5 <0.6")
-    cachy2.requires = [msgpack_dep]
+    cachy2.add_dependency(msgpack_dep)
 
     repo.add_package(get_package("cachy", "0.1.0"))
     repo.add_package(cachy2)
@@ -953,12 +1294,17 @@ Package operations: 2 installs, 0 updates, 0 removals
   - Installing cachy (0.2.0)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 2
 
 
-def test_add_git_constraint_old_installer(app, repo, installer, old_tester):
+def test_add_git_constraint_old_installer(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
+):
     repo.add_package(get_package("pendulum", "1.4.4"))
     repo.add_package(get_package("cleo", "0.6.5"))
 
@@ -977,7 +1323,7 @@ Package operations: 2 installs, 0 updates, 0 removals
   - Installing demo (0.1.2 9cf87a2)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 2
 
@@ -989,7 +1335,12 @@ Package operations: 2 installs, 0 updates, 0 removals
     }
 
 
-def test_add_git_constraint_with_poetry_old_installer(app, repo, installer, old_tester):
+def test_add_git_constraint_with_poetry_old_installer(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
+):
     repo.add_package(get_package("pendulum", "1.4.4"))
 
     old_tester.execute("git+https://github.com/demo/pyproject-demo.git")
@@ -1007,17 +1358,24 @@ Package operations: 2 installs, 0 updates, 0 removals
   - Installing demo (0.1.2 9cf87a2)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 2
 
 
-def test_add_git_constraint_with_extras_old_installer(app, repo, installer, old_tester):
+@pytest.mark.parametrize("extra_name", ["foo", "FOO"])
+def test_add_git_constraint_with_extras_old_installer(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
+    extra_name: str,
+):
     repo.add_package(get_package("pendulum", "1.4.4"))
     repo.add_package(get_package("cleo", "0.6.5"))
     repo.add_package(get_package("tomlkit", "0.5.5"))
 
-    old_tester.execute("git+https://github.com/demo/demo.git[foo,bar]")
+    old_tester.execute(f"git+https://github.com/demo/demo.git[{extra_name},bar]")
 
     expected = """\
 
@@ -1034,7 +1392,7 @@ Package operations: 4 installs, 0 updates, 0 removals
   - Installing demo (0.1.2 9cf87a2)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 4
 
@@ -1043,11 +1401,16 @@ Package operations: 4 installs, 0 updates, 0 removals
     assert "demo" in content["dependencies"]
     assert content["dependencies"]["demo"] == {
         "git": "https://github.com/demo/demo.git",
-        "extras": ["foo", "bar"],
+        "extras": [extra_name, "bar"],
     }
 
 
-def test_add_git_ssh_constraint_old_installer(app, repo, installer, old_tester):
+def test_add_git_ssh_constraint_old_installer(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
+):
     repo.add_package(get_package("pendulum", "1.4.4"))
     repo.add_package(get_package("cleo", "0.6.5"))
 
@@ -1066,7 +1429,7 @@ Package operations: 2 installs, 0 updates, 0 removals
   - Installing demo (0.1.2 9cf87a2)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 2
 
@@ -1079,19 +1442,23 @@ Package operations: 2 installs, 0 updates, 0 removals
     }
 
 
+@pytest.mark.parametrize(
+    "required_fixtures",
+    [["git/github.com/demo/demo"]],
+)
 def test_add_directory_constraint_old_installer(
-    app, repo, installer, mocker, old_tester
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
 ):
-    p = mocker.patch("poetry.utils._compat.Path.cwd")
-    p.return_value = Path(__file__) / ".."
-
     repo.add_package(get_package("pendulum", "1.4.4"))
     repo.add_package(get_package("cleo", "0.6.5"))
 
     path = "../git/github.com/demo/demo"
-    old_tester.execute("{}".format(path))
+    old_tester.execute(f"{path}")
 
-    expected = """\
+    expected = f"""\
 
 Updating dependencies
 Resolving dependencies...
@@ -1101,33 +1468,35 @@ Writing lock file
 Package operations: 2 installs, 0 updates, 0 removals
 
   - Installing pendulum (1.4.4)
-  - Installing demo (0.1.2 {})
-""".format(
-        app.poetry.file.parent.joinpath(path).resolve().as_posix()
-    )
+  - Installing demo (0.1.2 {app.poetry.file.parent.joinpath(path).resolve().as_posix()})
+"""
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 2
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
     assert "demo" in content["dependencies"]
-    assert content["dependencies"]["demo"] == {"path": "../git/github.com/demo/demo"}
+    assert content["dependencies"]["demo"] == {"path": path}
 
 
+@pytest.mark.parametrize(
+    "required_fixtures",
+    [["git/github.com/demo/pyproject-demo"]],
+)
 def test_add_directory_with_poetry_old_installer(
-    app, repo, installer, mocker, old_tester
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
 ):
-    p = mocker.patch("poetry.utils._compat.Path.cwd")
-    p.return_value = Path(__file__) / ".."
-
     repo.add_package(get_package("pendulum", "1.4.4"))
 
     path = "../git/github.com/demo/pyproject-demo"
-    old_tester.execute("{}".format(path))
+    old_tester.execute(f"{path}")
 
-    expected = """\
+    expected = f"""\
 
 Updating dependencies
 Resolving dependencies...
@@ -1137,28 +1506,30 @@ Writing lock file
 Package operations: 2 installs, 0 updates, 0 removals
 
   - Installing pendulum (1.4.4)
-  - Installing demo (0.1.2 {})
-""".format(
-        app.poetry.file.parent.joinpath(path).resolve().as_posix()
-    )
+  - Installing demo (0.1.2 {app.poetry.file.parent.joinpath(path).resolve().as_posix()})
+"""
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 2
 
 
+@pytest.mark.parametrize(
+    "required_fixtures",
+    [["distributions/demo-0.1.0-py2.py3-none-any.whl"]],
+)
 def test_add_file_constraint_wheel_old_installer(
-    app, repo, installer, mocker, old_tester
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
 ):
-    p = mocker.patch("poetry.utils._compat.Path.cwd")
-    p.return_value = Path(__file__) / ".."
-
     repo.add_package(get_package("pendulum", "1.4.4"))
 
     path = "../distributions/demo-0.1.0-py2.py3-none-any.whl"
-    old_tester.execute("{}".format(path))
+    old_tester.execute(f"{path}")
 
-    expected = """\
+    expected = f"""\
 
 Updating dependencies
 Resolving dependencies...
@@ -1168,35 +1539,35 @@ Writing lock file
 Package operations: 2 installs, 0 updates, 0 removals
 
   - Installing pendulum (1.4.4)
-  - Installing demo (0.1.0 {})
-""".format(
-        app.poetry.file.parent.joinpath(path).resolve().as_posix()
-    )
+  - Installing demo (0.1.0 {app.poetry.file.parent.joinpath(path).resolve().as_posix()})
+"""
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 2
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
     assert "demo" in content["dependencies"]
-    assert content["dependencies"]["demo"] == {
-        "path": "../distributions/demo-0.1.0-py2.py3-none-any.whl"
-    }
+    assert content["dependencies"]["demo"] == {"path": path}
 
 
+@pytest.mark.parametrize(
+    "required_fixtures",
+    [["distributions/demo-0.1.0.tar.gz"]],
+)
 def test_add_file_constraint_sdist_old_installer(
-    app, repo, installer, mocker, old_tester
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
 ):
-    p = mocker.patch("poetry.utils._compat.Path.cwd")
-    p.return_value = Path(__file__) / ".."
-
     repo.add_package(get_package("pendulum", "1.4.4"))
 
     path = "../distributions/demo-0.1.0.tar.gz"
-    old_tester.execute("{}".format(path))
+    old_tester.execute(f"{path}")
 
-    expected = """\
+    expected = f"""\
 
 Updating dependencies
 Resolving dependencies...
@@ -1206,36 +1577,37 @@ Writing lock file
 Package operations: 2 installs, 0 updates, 0 removals
 
   - Installing pendulum (1.4.4)
-  - Installing demo (0.1.0 {})
-""".format(
-        app.poetry.file.parent.joinpath(path).resolve().as_posix()
-    )
+  - Installing demo (0.1.0 {app.poetry.file.parent.joinpath(path).resolve().as_posix()})
+"""
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 2
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
     assert "demo" in content["dependencies"]
-    assert content["dependencies"]["demo"] == {
-        "path": "../distributions/demo-0.1.0.tar.gz"
-    }
+    assert content["dependencies"]["demo"] == {"path": path}
 
 
+@pytest.mark.parametrize("extra_name", ["msgpack", "MsgPack"])
 def test_add_constraint_with_extras_option_old_installer(
-    app, repo, installer, old_tester
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
+    extra_name: str,
 ):
     cachy2 = get_package("cachy", "0.2.0")
     cachy2.extras = {"msgpack": [get_dependency("msgpack-python")]}
     msgpack_dep = get_dependency("msgpack-python", ">=0.5 <0.6", optional=True)
-    cachy2.requires = [msgpack_dep]
+    cachy2.add_dependency(msgpack_dep)
 
     repo.add_package(get_package("cachy", "0.1.0"))
     repo.add_package(cachy2)
     repo.add_package(get_package("msgpack-python", "0.5.3"))
 
-    old_tester.execute("cachy=0.2.0 --extras msgpack")
+    old_tester.execute(f"cachy=0.2.0 --extras {extra_name}")
 
     expected = """\
 
@@ -1250,7 +1622,7 @@ Package operations: 2 installs, 0 updates, 0 removals
   - Installing cachy (0.2.0)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 2
 
@@ -1259,14 +1631,18 @@ Package operations: 2 installs, 0 updates, 0 removals
     assert "cachy" in content["dependencies"]
     assert content["dependencies"]["cachy"] == {
         "version": "0.2.0",
-        "extras": ["msgpack"],
+        "extras": [extra_name],
     }
 
 
 def test_add_url_constraint_wheel_old_installer(
-    app, repo, installer, mocker, old_tester
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    mocker: MockerFixture,
+    old_tester: CommandTester,
 ):
-    p = mocker.patch("poetry.utils._compat.Path.cwd")
+    p = mocker.patch("pathlib.Path.cwd")
     p.return_value = Path(__file__) / ".."
 
     repo.add_package(get_package("pendulum", "1.4.4"))
@@ -1285,10 +1661,11 @@ Writing lock file
 Package operations: 2 installs, 0 updates, 0 removals
 
   - Installing pendulum (1.4.4)
-  - Installing demo (0.1.0 https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl)
+  - Installing demo\
+ (0.1.0 https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 2
 
@@ -1300,15 +1677,21 @@ Package operations: 2 installs, 0 updates, 0 removals
     }
 
 
+@pytest.mark.parametrize("extra_name", ["foo", "FOO"])
 def test_add_url_constraint_wheel_with_extras_old_installer(
-    app, repo, installer, old_tester
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
+    extra_name: str,
 ):
     repo.add_package(get_package("pendulum", "1.4.4"))
     repo.add_package(get_package("cleo", "0.6.5"))
     repo.add_package(get_package("tomlkit", "0.5.5"))
 
     old_tester.execute(
-        "https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl[foo,bar]"
+        "https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl"
+        f"[{extra_name},bar]"
     )
 
     expected = """\
@@ -1323,10 +1706,11 @@ Package operations: 4 installs, 0 updates, 0 removals
   - Installing cleo (0.6.5)
   - Installing pendulum (1.4.4)
   - Installing tomlkit (0.5.5)
-  - Installing demo (0.1.0 https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl)
+  - Installing demo\
+ (0.1.0 https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 4
 
@@ -1334,12 +1718,19 @@ Package operations: 4 installs, 0 updates, 0 removals
 
     assert "demo" in content["dependencies"]
     assert content["dependencies"]["demo"] == {
-        "url": "https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl",
-        "extras": ["foo", "bar"],
+        "url": (
+            "https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl"
+        ),
+        "extras": [extra_name, "bar"],
     }
 
 
-def test_add_constraint_with_python_old_installer(app, repo, installer, old_tester):
+def test_add_constraint_with_python_old_installer(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
+):
     cachy2 = get_package("cachy", "0.2.0")
 
     repo.add_package(get_package("cachy", "0.1.0"))
@@ -1359,7 +1750,7 @@ Package operations: 1 install, 0 updates, 0 removals
   - Installing cachy (0.2.0)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 1
 
@@ -1370,7 +1761,11 @@ Package operations: 1 install, 0 updates, 0 removals
 
 
 def test_add_constraint_with_platform_old_installer(
-    app, repo, installer, env, old_tester
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    env: MockEnv,
+    old_tester: CommandTester,
 ):
     platform = sys.platform
     env._platform = platform
@@ -1380,7 +1775,7 @@ def test_add_constraint_with_platform_old_installer(
     repo.add_package(get_package("cachy", "0.1.0"))
     repo.add_package(cachy2)
 
-    old_tester.execute("cachy=0.2.0 --platform {} -vvv".format(platform))
+    old_tester.execute(f"cachy=0.2.0 --platform {platform} -vvv")
 
     expected = """\
 
@@ -1394,7 +1789,7 @@ Package operations: 1 install, 0 updates, 0 removals
   - Installing cachy (0.2.0)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 1
 
@@ -1407,7 +1802,12 @@ Package operations: 1 install, 0 updates, 0 removals
     }
 
 
-def test_add_constraint_with_source_old_installer(app, poetry, installer, old_tester):
+def test_add_constraint_with_source_old_installer(
+    app: PoetryTestApplication,
+    poetry: Poetry,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
+):
     repo = LegacyRepository(name="my-index", url="https://my-index.fake")
     repo.add_package(get_package("cachy", "0.2.0"))
     repo._cache.store("matches").put("cachy:0.2.0", [Version.parse("0.2.0")], 5)
@@ -1428,7 +1828,7 @@ Package operations: 1 install, 0 updates, 0 removals
   - Installing cachy (0.2.0)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 1
 
@@ -1441,15 +1841,20 @@ Package operations: 1 install, 0 updates, 0 removals
     }
 
 
-def test_add_constraint_with_source_that_does_not_exist_old_installer(app, old_tester):
+def test_add_constraint_with_source_that_does_not_exist_old_installer(
+    app: PoetryTestApplication, old_tester: CommandTester
+):
     with pytest.raises(ValueError) as e:
         old_tester.execute("foo --source i-dont-exist")
 
-    assert 'Repository "i-dont-exist" does not exist.' == str(e.value)
+    assert str(e.value) == 'Repository "i-dont-exist" does not exist.'
 
 
 def test_add_constraint_not_found_with_source_old_installer(
-    app, poetry, mocker, old_tester
+    app: PoetryTestApplication,
+    poetry: Poetry,
+    mocker: MockerFixture,
+    old_tester: CommandTester,
 ):
     repo = LegacyRepository(name="my-index", url="https://my-index.fake")
     mocker.patch.object(repo, "find_packages", return_value=[])
@@ -1462,16 +1867,19 @@ def test_add_constraint_not_found_with_source_old_installer(
     with pytest.raises(ValueError) as e:
         old_tester.execute("cachy --source my-index")
 
-    assert "Could not find a matching version of package cachy" == str(e.value)
+    assert str(e.value) == "Could not find a matching version of package cachy"
 
 
 def test_add_to_section_that_does_no_exist_yet_old_installer(
-    app, repo, installer, old_tester
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
 ):
     repo.add_package(get_package("cachy", "0.1.0"))
     repo.add_package(get_package("cachy", "0.2.0"))
 
-    old_tester.execute("cachy --dev")
+    old_tester.execute("cachy --group dev")
 
     expected = """\
 Using version ^0.2.0 for cachy
@@ -1486,18 +1894,21 @@ Package operations: 1 install, 0 updates, 0 removals
   - Installing cachy (0.2.0)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 1
 
     content = app.poetry.file.read()["tool"]["poetry"]
 
-    assert "cachy" in content["dev-dependencies"]
-    assert content["dev-dependencies"]["cachy"] == "^0.2.0"
+    assert "cachy" in content["group"]["dev"]["dependencies"]
+    assert content["group"]["dev"]["dependencies"]["cachy"] == "^0.2.0"
 
 
 def test_add_should_not_select_prereleases_old_installer(
-    app, repo, installer, old_tester
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
 ):
     repo.add_package(get_package("pyyaml", "3.13"))
     repo.add_package(get_package("pyyaml", "4.2b2"))
@@ -1517,7 +1928,7 @@ Package operations: 1 install, 0 updates, 0 removals
   - Installing pyyaml (3.13)
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
 
     assert len(installer.installs) == 1
 
@@ -1527,8 +1938,11 @@ Package operations: 1 install, 0 updates, 0 removals
     assert content["dependencies"]["pyyaml"] == "^3.13"
 
 
-def test_add_should_display_an_error_when_adding_existing_package_with_no_constraint_old_installer(
-    app, repo, installer, old_tester
+def test_add_should_skip_when_adding_existing_package_with_no_constraint_old_installer(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
 ):
     content = app.poetry.file.read()
     content["tool"]["poetry"]["dependencies"]["foo"] = "^1.0"
@@ -1536,14 +1950,27 @@ def test_add_should_display_an_error_when_adding_existing_package_with_no_constr
 
     repo.add_package(get_package("foo", "1.1.2"))
 
-    with pytest.raises(ValueError) as e:
-        old_tester.execute("foo")
+    old_tester.execute("foo")
 
-    assert "Package foo is already present" == str(e.value)
+    expected = """\
+The following packages are already present in the pyproject.toml and will be skipped:
+
+  • foo
+
+If you want to update it to the latest compatible version,\
+ you can use `poetry update package`.
+If you prefer to upgrade it to the latest available version,\
+ you can use `poetry add package@latest`.
+"""
+
+    assert expected in old_tester.io.fetch_output()
 
 
-def test_add_should_work_when_adding_existing_package_with_latest_constraint_old_installer(
-    app, repo, installer, old_tester
+def test_add_should_work_when_adding_existing_package_with_latest_constraint_old_installer(  # noqa: E501
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
 ):
     content = app.poetry.file.read()
     content["tool"]["poetry"]["dependencies"]["foo"] = "^1.0"
@@ -1575,7 +2002,10 @@ Package operations: 1 install, 0 updates, 0 removals
 
 
 def test_add_chooses_prerelease_if_only_prereleases_are_available_old_installer(
-    app, repo, installer, old_tester
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
 ):
     repo.add_package(get_package("foo", "1.2.3b0"))
     repo.add_package(get_package("foo", "1.2.3b1"))
@@ -1583,7 +2013,7 @@ def test_add_chooses_prerelease_if_only_prereleases_are_available_old_installer(
     old_tester.execute("foo")
 
     expected = """\
-Using version ^1.2.3-beta.1 for foo
+Using version ^1.2.3b1 for foo
 
 Updating dependencies
 Resolving dependencies...
@@ -1594,11 +2024,15 @@ Package operations: 1 install, 0 updates, 0 removals
 
   - Installing foo (1.2.3b1)
 """
-
     assert expected in old_tester.io.fetch_output()
 
 
-def test_add_preferes_stable_releases_old_installer(app, repo, installer, old_tester):
+def test_add_preferes_stable_releases_old_installer(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
+):
     repo.add_package(get_package("foo", "1.2.3"))
     repo.add_package(get_package("foo", "1.2.4b1"))
 
@@ -1620,7 +2054,12 @@ Package operations: 1 install, 0 updates, 0 removals
     assert expected in old_tester.io.fetch_output()
 
 
-def test_add_with_lock_old_installer(app, repo, installer, old_tester):
+def test_add_with_lock_old_installer(
+    app: PoetryTestApplication,
+    repo: TestRepository,
+    installer: NoopInstaller,
+    old_tester: CommandTester,
+):
     repo.add_package(get_package("cachy", "0.2.0"))
 
     old_tester.execute("cachy --lock")
@@ -1634,4 +2073,58 @@ Resolving dependencies...
 Writing lock file
 """
 
-    assert expected == old_tester.io.fetch_output()
+    assert old_tester.io.fetch_output() == expected
+
+
+def test_add_keyboard_interrupt_restore_content(
+    poetry_with_up_to_date_lockfile: Poetry,
+    repo: TestRepository,
+    command_tester_factory: CommandTesterFactory,
+    mocker: MockerFixture,
+):
+    tester = command_tester_factory("add", poetry=poetry_with_up_to_date_lockfile)
+
+    mocker.patch(
+        "poetry.installation.installer.Installer.run", side_effect=KeyboardInterrupt()
+    )
+    original_pyproject_content = poetry_with_up_to_date_lockfile.file.read()
+    original_lockfile_content = poetry_with_up_to_date_lockfile._locker.lock_data
+
+    repo.add_package(get_package("cachy", "0.2.0"))
+    repo.add_package(get_package("docker", "4.3.1"))
+
+    tester.execute("cachy")
+
+    assert poetry_with_up_to_date_lockfile.file.read() == original_pyproject_content
+    assert (
+        poetry_with_up_to_date_lockfile._locker.lock_data == original_lockfile_content
+    )
+
+
+@pytest.mark.parametrize(
+    "command",
+    [
+        "cachy --dry-run",
+        "cachy --lock --dry-run",
+    ],
+)
+def test_add_with_dry_run_keep_files_intact(
+    command: str,
+    poetry_with_up_to_date_lockfile: Poetry,
+    repo: TestRepository,
+    command_tester_factory: CommandTesterFactory,
+):
+    tester = command_tester_factory("add", poetry=poetry_with_up_to_date_lockfile)
+
+    original_pyproject_content = poetry_with_up_to_date_lockfile.file.read()
+    original_lockfile_content = poetry_with_up_to_date_lockfile._locker.lock_data
+
+    repo.add_package(get_package("cachy", "0.2.0"))
+    repo.add_package(get_package("docker", "4.3.1"))
+
+    tester.execute(command)
+
+    assert poetry_with_up_to_date_lockfile.file.read() == original_pyproject_content
+    assert (
+        poetry_with_up_to_date_lockfile._locker.lock_data == original_lockfile_content
+    )
