@@ -1,20 +1,39 @@
-from cleo.testers import CommandTester
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+import pytest
 
 from poetry.factory import Factory
 from tests.helpers import get_package
 
 
-def test_debug_resolve_gives_resolution_results(app, repo):
-    command = app.find("debug resolve")
-    tester = CommandTester(command)
+if TYPE_CHECKING:
+    from cleo.testers.command_tester import CommandTester
 
-    cachy2 = get_package("cachy", "0.2.0")
-    cachy2.add_dependency(Factory.create_dependency("msgpack-python", ">=0.5 <0.6"))
+    from tests.helpers import TestRepository
+    from tests.types import CommandTesterFactory
+
+
+@pytest.fixture()
+def tester(command_tester_factory: CommandTesterFactory) -> CommandTester:
+    return command_tester_factory("debug resolve")
+
+
+@pytest.fixture(autouse=True)
+def __add_packages(repo: TestRepository) -> None:
+    cachy020 = get_package("cachy", "0.2.0")
+    cachy020.add_dependency(Factory.create_dependency("msgpack-python", ">=0.5 <0.6"))
 
     repo.add_package(get_package("cachy", "0.1.0"))
-    repo.add_package(cachy2)
+    repo.add_package(cachy020)
     repo.add_package(get_package("msgpack-python", "0.5.3"))
 
+    repo.add_package(get_package("pendulum", "2.0.3"))
+    repo.add_package(get_package("cleo", "0.6.5"))
+
+
+def test_debug_resolve_gives_resolution_results(tester: CommandTester):
     tester.execute("cachy")
 
     expected = """\
@@ -26,20 +45,10 @@ msgpack-python 0.5.3
 cachy          0.2.0
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_debug_resolve_tree_option_gives_the_dependency_tree(app, repo):
-    command = app.find("debug resolve")
-    tester = CommandTester(command)
-
-    cachy2 = get_package("cachy", "0.2.0")
-    cachy2.add_dependency(Factory.create_dependency("msgpack-python", ">=0.5 <0.6"))
-
-    repo.add_package(get_package("cachy", "0.1.0"))
-    repo.add_package(cachy2)
-    repo.add_package(get_package("msgpack-python", "0.5.3"))
-
+def test_debug_resolve_tree_option_gives_the_dependency_tree(tester: CommandTester):
     tester.execute("cachy --tree")
 
     expected = """\
@@ -48,19 +57,13 @@ Resolving dependencies...
 Resolution results:
 
 cachy 0.2.0
-`-- msgpack-python >=0.5 <0.6
+└── msgpack-python >=0.5 <0.6
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_debug_resolve_git_dependency(app, repo):
-    repo.add_package(get_package("pendulum", "2.0.3"))
-    repo.add_package(get_package("cleo", "0.6.5"))
-
-    command = app.find("debug resolve")
-    tester = CommandTester(command)
-
+def test_debug_resolve_git_dependency(tester: CommandTester):
     tester.execute("git+https://github.com/demo/demo.git")
 
     expected = """\
@@ -72,4 +75,4 @@ pendulum 2.0.3
 demo     0.1.2
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
