@@ -1,12 +1,39 @@
-from packaging.tags import Tag
+from __future__ import annotations
 
+from pathlib import Path
+from typing import TYPE_CHECKING
+
+import pytest
+
+from packaging.tags import Tag
 from poetry.core.packages.utils.link import Link
+
 from poetry.installation.chef import Chef
-from poetry.utils._compat import Path
 from poetry.utils.env import MockEnv
 
 
-def test_get_cached_archive_for_link(config, mocker):
+if TYPE_CHECKING:
+    from pytest_mock import MockerFixture
+
+    from tests.conftest import Config
+
+
+@pytest.mark.parametrize(
+    ("link", "cached"),
+    [
+        (
+            "https://files.python-poetry.org/demo-0.1.0.tar.gz",
+            "/cache/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl",
+        ),
+        (
+            "https://example.com/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl",
+            "/cache/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl",
+        ),
+    ],
+)
+def test_get_cached_archive_for_link(
+    config: Config, mocker: MockerFixture, link: str, cached: str
+):
     chef = Chef(
         config,
         MockEnv(
@@ -23,21 +50,19 @@ def test_get_cached_archive_for_link(config, mocker):
         chef,
         "get_cached_archives_for_link",
         return_value=[
-            Link("file:///foo/demo-0.1.0-py2.py3-none-any"),
-            Link("file:///foo/demo-0.1.0.tar.gz"),
-            Link("file:///foo/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl"),
-            Link("file:///foo/demo-0.1.0-cp37-cp37-macosx_10_15_x86_64.whl"),
+            Path("/cache/demo-0.1.0-py2.py3-none-any"),
+            Path("/cache/demo-0.1.0.tar.gz"),
+            Path("/cache/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl"),
+            Path("/cache/demo-0.1.0-cp37-cp37-macosx_10_15_x86_64.whl"),
         ],
     )
 
-    archive = chef.get_cached_archive_for_link(
-        Link("https://files.python-poetry.org/demo-0.1.0.tar.gz")
-    )
+    archive = chef.get_cached_archive_for_link(Link(link))
 
-    assert Link("file:///foo/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl") == archive
+    assert Path(cached) == archive
 
 
-def test_get_cached_archives_for_link(config, mocker):
+def test_get_cached_archives_for_link(config: Config, mocker: MockerFixture):
     chef = Chef(
         config,
         MockEnv(
@@ -47,7 +72,9 @@ def test_get_cached_archives_for_link(config, mocker):
 
     distributions = Path(__file__).parent.parent.joinpath("fixtures/distributions")
     mocker.patch.object(
-        chef, "get_cache_directory_for_link", return_value=distributions,
+        chef,
+        "get_cache_directory_for_link",
+        return_value=distributions,
     )
 
     archives = chef.get_cached_archives_for_link(
@@ -55,12 +82,10 @@ def test_get_cached_archives_for_link(config, mocker):
     )
 
     assert archives
-    assert set(archives) == {
-        Link(path.as_uri()) for path in distributions.glob("demo-0.1.0*")
-    }
+    assert set(archives) == {Path(path) for path in distributions.glob("demo-0.1.0*")}
 
 
-def test_get_cache_directory_for_link(config):
+def test_get_cache_directory_for_link(config: Config, config_cache_dir: Path):
     chef = Chef(
         config,
         MockEnv(
@@ -71,8 +96,10 @@ def test_get_cache_directory_for_link(config):
     directory = chef.get_cache_directory_for_link(
         Link("https://files.python-poetry.org/poetry-1.1.0.tar.gz")
     )
+
     expected = Path(
-        "/foo/artifacts/ba/63/13/283a3b3b7f95f05e9e6f84182d276f7bb0951d5b0cc24422b33f7a4648"
+        f"{config_cache_dir.as_posix()}/artifacts/ba/63/13/"
+        "283a3b3b7f95f05e9e6f84182d276f7bb0951d5b0cc24422b33f7a4648"
     )
 
-    assert expected == directory
+    assert directory == expected
