@@ -1,10 +1,22 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
 from poetry.factory import Factory
+from tests.mixology.helpers import add_to_repo
+from tests.mixology.helpers import check_solver_result
 
-from ..helpers import add_to_repo
-from ..helpers import check_solver_result
+
+if TYPE_CHECKING:
+    from poetry.core.packages.project_package import ProjectPackage
+
+    from poetry.repositories import Repository
+    from tests.mixology.version_solver.conftest import Provider
 
 
-def test_circular_dependency_on_older_version(root, provider, repo):
+def test_circular_dependency_on_older_version(
+    root: ProjectPackage, provider: Provider, repo: Repository
+):
     root.add_dependency(Factory.create_dependency("a", ">=1.0.0"))
 
     add_to_repo(repo, "a", "1.0.0")
@@ -14,7 +26,9 @@ def test_circular_dependency_on_older_version(root, provider, repo):
     check_solver_result(root, provider, {"a": "1.0.0"}, tries=2)
 
 
-def test_diamond_dependency_graph(root, provider, repo):
+def test_diamond_dependency_graph(
+    root: ProjectPackage, provider: Provider, repo: Repository
+):
     root.add_dependency(Factory.create_dependency("a", "*"))
     root.add_dependency(Factory.create_dependency("b", "*"))
 
@@ -31,7 +45,9 @@ def test_diamond_dependency_graph(root, provider, repo):
     check_solver_result(root, provider, {"a": "1.0.0", "b": "2.0.0", "c": "3.0.0"})
 
 
-def test_backjumps_after_partial_satisfier(root, provider, repo):
+def test_backjumps_after_partial_satisfier(
+    root: ProjectPackage, provider: Provider, repo: Repository
+):
     # c 2.0.0 is incompatible with y 2.0.0 because it requires x 1.0.0, but that
     # requirement only exists because of both a and b. The solver should be able
     # to deduce c 2.0.0's incompatibility and select c 1.0.0 instead.
@@ -54,7 +70,9 @@ def test_backjumps_after_partial_satisfier(root, provider, repo):
     check_solver_result(root, provider, {"c": "1.0.0", "y": "2.0.0"}, tries=2)
 
 
-def test_rolls_back_leaf_versions_first(root, provider, repo):
+def test_rolls_back_leaf_versions_first(
+    root: ProjectPackage, provider: Provider, repo: Repository
+):
     # The latest versions of a and b disagree on c. An older version of either
     # will resolve the problem. This test validates that b, which is farther
     # in the dependency graph from myapp is downgraded first.
@@ -70,7 +88,7 @@ def test_rolls_back_leaf_versions_first(root, provider, repo):
     check_solver_result(root, provider, {"a": "2.0.0", "b": "1.0.0", "c": "2.0.0"})
 
 
-def test_simple_transitive(root, provider, repo):
+def test_simple_transitive(root: ProjectPackage, provider: Provider, repo: Repository):
     # Only one version of baz, so foo and bar will have to downgrade
     # until they reach it
     root.add_dependency(Factory.create_dependency("foo", "*"))
@@ -90,7 +108,9 @@ def test_simple_transitive(root, provider, repo):
     )
 
 
-def test_backjump_to_nearer_unsatisfied_package(root, provider, repo):
+def test_backjump_to_nearer_unsatisfied_package(
+    root: ProjectPackage, provider: Provider, repo: Repository
+):
     # This ensures it doesn't exhaustively search all versions of b when it's
     # a-2.0.0 whose dependency on c-2.0.0-nonexistent led to the problem. We
     # make sure b has more versions than a so that the solver tries a first
@@ -99,7 +119,7 @@ def test_backjump_to_nearer_unsatisfied_package(root, provider, repo):
     root.add_dependency(Factory.create_dependency("b", "*"))
 
     add_to_repo(repo, "a", "1.0.0", deps={"c": "1.0.0"})
-    add_to_repo(repo, "a", "2.0.0", deps={"c": "2.0.0-nonexistent"})
+    add_to_repo(repo, "a", "2.0.0", deps={"c": "2.0.0-1"})
     add_to_repo(repo, "b", "1.0.0")
     add_to_repo(repo, "b", "2.0.0")
     add_to_repo(repo, "b", "3.0.0")
@@ -110,7 +130,9 @@ def test_backjump_to_nearer_unsatisfied_package(root, provider, repo):
     )
 
 
-def test_traverse_into_package_with_fewer_versions_first(root, provider, repo):
+def test_traverse_into_package_with_fewer_versions_first(
+    root: ProjectPackage, provider: Provider, repo: Repository
+):
     # Dependencies are ordered so that packages with fewer versions are tried
     # first. Here, there are two valid solutions (either a or b must be
     # downgraded once). The chosen one depends on which dep is traversed first.
@@ -134,7 +156,9 @@ def test_traverse_into_package_with_fewer_versions_first(root, provider, repo):
     check_solver_result(root, provider, {"a": "4.0.0", "b": "4.0.0", "c": "2.0.0"})
 
 
-def test_backjump_past_failed_package_on_disjoint_constraint(root, provider, repo):
+def test_backjump_past_failed_package_on_disjoint_constraint(
+    root: ProjectPackage, provider: Provider, repo: Repository
+):
     root.add_dependency(Factory.create_dependency("a", "*"))
     root.add_dependency(Factory.create_dependency("foo", ">2.0.0"))
 
