@@ -1,41 +1,55 @@
+from __future__ import annotations
+
+from pathlib import Path
 from subprocess import CalledProcessError
+from typing import TYPE_CHECKING
 
 import pytest
 
-from clikit.io import NullIO
-
-from poetry.core.packages import ProjectPackage
+from cleo.io.null_io import NullIO
+from poetry.core.packages.dependency import Dependency
 from poetry.core.packages.directory_dependency import DirectoryDependency
 from poetry.core.packages.file_dependency import FileDependency
+from poetry.core.packages.package import Package
+from poetry.core.packages.project_package import ProjectPackage
+from poetry.core.packages.url_dependency import URLDependency
 from poetry.core.packages.vcs_dependency import VCSDependency
+
+from poetry.factory import Factory
 from poetry.inspection.info import PackageInfo
+from poetry.packages import DependencyPackage
 from poetry.puzzle.provider import Provider
 from poetry.repositories.pool import Pool
 from poetry.repositories.repository import Repository
-from poetry.utils._compat import PY35
-from poetry.utils._compat import Path
 from poetry.utils.env import EnvCommandError
 from poetry.utils.env import MockEnv as BaseMockEnv
 from tests.helpers import get_dependency
 
 
+if TYPE_CHECKING:
+    from pytest_mock import MockerFixture
+
+
+SOME_URL = "https://example.com/path.tar.gz"
+
+
 class MockEnv(BaseMockEnv):
-    def run(self, bin, *args):
+    def run(self, bin: str, *args: str) -> None:
         raise EnvCommandError(CalledProcessError(1, "python", output=""))
 
 
 @pytest.fixture
-def root():
+def root() -> ProjectPackage:
     return ProjectPackage("root", "1.2.3")
 
 
 @pytest.fixture
-def repository():
-    return Repository()
+def repository() -> Repository:
+    return Repository("repo")
 
 
 @pytest.fixture
-def pool(repository):
+def pool(repository: Repository) -> Pool:
     pool = Pool()
     pool.add_repository(repository)
 
@@ -43,102 +57,215 @@ def pool(repository):
 
 
 @pytest.fixture
-def provider(root, pool):
+def provider(root: ProjectPackage, pool: Pool) -> Provider:
     return Provider(root, pool, NullIO())
 
 
-def test_search_for_vcs_setup_egg_info(provider):
-    dependency = VCSDependency("demo", "git", "https://github.com/demo/demo.git")
-
-    package = provider.search_for_vcs(dependency)[0]
-
-    assert package.name == "demo"
-    assert package.version.text == "0.1.2"
-
-    required = [r for r in package.requires if not r.is_optional()]
-    optional = [r for r in package.requires if r.is_optional()]
-    assert required == [get_dependency("pendulum", ">=1.4.4")]
-    assert optional == [get_dependency("tomlkit"), get_dependency("cleo")]
-    assert package.extras == {
-        "foo": [get_dependency("cleo")],
-        "bar": [get_dependency("tomlkit")],
-    }
-
-
-def test_search_for_vcs_setup_egg_info_with_extras(provider):
-    dependency = VCSDependency(
-        "demo", "git", "https://github.com/demo/demo.git", extras=["foo"]
-    )
-
-    package = provider.search_for_vcs(dependency)[0]
-
-    assert package.name == "demo"
-    assert package.version.text == "0.1.2"
-
-    required = [r for r in package.requires if not r.is_optional()]
-    optional = [r for r in package.requires if r.is_optional()]
-    assert required == [get_dependency("pendulum", ">=1.4.4")]
-    assert optional == [get_dependency("tomlkit"), get_dependency("cleo")]
-    assert package.extras == {
-        "foo": [get_dependency("cleo")],
-        "bar": [get_dependency("tomlkit")],
-    }
-
-
-@pytest.mark.skipif(not PY35, reason="AST parsing does not work for Python <3.4")
-def test_search_for_vcs_read_setup(provider, mocker):
-    mocker.patch("poetry.utils.env.EnvManager.get", return_value=MockEnv())
-
-    dependency = VCSDependency("demo", "git", "https://github.com/demo/demo.git")
-
-    package = provider.search_for_vcs(dependency)[0]
-
-    assert package.name == "demo"
-    assert package.version.text == "0.1.2"
-
-    required = [r for r in package.requires if not r.is_optional()]
-    optional = [r for r in package.requires if r.is_optional()]
-    assert required == [get_dependency("pendulum", ">=1.4.4")]
-    assert optional == [get_dependency("tomlkit"), get_dependency("cleo")]
-    assert package.extras == {
-        "foo": [get_dependency("cleo")],
-        "bar": [get_dependency("tomlkit")],
-    }
+@pytest.mark.parametrize(
+    "dependency, expected",
+    [
+        (Dependency("foo", "<2"), [Package("foo", "1")]),
+        (Dependency("foo", "<2", extras=["bar"]), [Package("foo", "1")]),
+        (Dependency("foo", ">=1"), [Package("foo", "2"), Package("foo", "1")]),
+        (
+            Dependency("foo", ">=1a"),
+            [
+                Package("foo", "3a"),
+                Package("foo", "2"),
+                Package("foo", "2a"),
+                Package("foo", "1"),
+            ],
+        ),
+        (
+            Dependency("foo", ">=1", allows_prereleases=True),
+            [
+                Package("foo", "3a"),
+                Package("foo", "2"),
+                Package("foo", "2a"),
+                Package("foo", "1"),
+            ],
+        ),
+    ],
+)
+def test_search_for(
+    provider: Provider,
+    repository: Repository,
+    dependency: Dependency,
+    expected: list[Package],
+) -> None:
+    foo1 = Package("foo", "1")
+    foo2a = Package("foo", "2a")
+    foo2 = Package("foo", "2")
+    foo3a = Package("foo", "3a")
+    repository.add_package(foo1)
+    repository.add_package(foo2a)
+    repository.add_package(foo2)
+    repository.add_package(foo3a)
+    assert provider.search_for(dependency) == expected
 
 
-@pytest.mark.skipif(not PY35, reason="AST parsing does not work for Python <3.4")
-def test_search_for_vcs_read_setup_with_extras(provider, mocker):
-    mocker.patch("poetry.utils.env.EnvManager.get", return_value=MockEnv())
-
-    dependency = VCSDependency(
-        "demo", "git", "https://github.com/demo/demo.git", extras=["foo"]
-    )
-
-    package = provider.search_for_vcs(dependency)[0]
-
-    assert package.name == "demo"
-    assert package.version.text == "0.1.2"
-
-    required = [r for r in package.requires if not r.is_optional()]
-    optional = [r for r in package.requires if r.is_optional()]
-    assert required == [get_dependency("pendulum", ">=1.4.4")]
-    assert optional == [get_dependency("tomlkit"), get_dependency("cleo")]
-
-
-def test_search_for_vcs_read_setup_raises_error_if_no_version(provider, mocker):
+@pytest.mark.parametrize(
+    "dependency, direct_origin_dependency, expected_before, expected_after",
+    [
+        (
+            Dependency("foo", ">=1"),
+            URLDependency("foo", SOME_URL),
+            [Package("foo", "3")],
+            [Package("foo", "2a", source_type="url", source_url=SOME_URL)],
+        ),
+        (
+            Dependency("foo", ">=2"),
+            URLDependency("foo", SOME_URL),
+            [Package("foo", "3")],
+            [],
+        ),
+        (
+            Dependency("foo", ">=1", extras=["bar"]),
+            URLDependency("foo", SOME_URL),
+            [Package("foo", "3")],
+            [Package("foo", "2a", source_type="url", source_url=SOME_URL)],
+        ),
+        (
+            Dependency("foo", ">=1"),
+            URLDependency("foo", SOME_URL, extras=["baz"]),
+            [Package("foo", "3")],
+            [Package("foo", "2a", source_type="url", source_url=SOME_URL)],
+        ),
+        (
+            Dependency("foo", ">=1", extras=["bar"]),
+            URLDependency("foo", SOME_URL, extras=["baz"]),
+            [Package("foo", "3")],
+            [Package("foo", "2a", source_type="url", source_url=SOME_URL)],
+        ),
+    ],
+)
+def test_search_for_direct_origin_and_extras(
+    provider: Provider,
+    repository: Repository,
+    mocker: MockerFixture,
+    dependency: Dependency,
+    direct_origin_dependency: Dependency,
+    expected_before: list[Package],
+    expected_after: list[Package],
+) -> None:
+    foo2a_direct_origin = Package("foo", "2a", source_type="url", source_url=SOME_URL)
     mocker.patch(
-        "poetry.inspection.info.PackageInfo._pep517_metadata",
+        "poetry.puzzle.provider.Provider.search_for_direct_origin_dependency",
+        return_value=foo2a_direct_origin,
+    )
+    foo2a = Package("foo", "2a")
+    foo3 = Package("foo", "3")
+    repository.add_package(foo2a)
+    repository.add_package(foo3)
+
+    assert provider.search_for(dependency) == expected_before
+    assert provider.search_for(direct_origin_dependency) == [foo2a_direct_origin]
+    assert provider.search_for(dependency) == expected_after
+
+
+@pytest.mark.parametrize("value", [True, False])
+def test_search_for_vcs_retains_develop_flag(provider: Provider, value: bool):
+    dependency = VCSDependency(
+        "demo", "git", "https://github.com/demo/demo.git", develop=value
+    )
+    package = provider.search_for_direct_origin_dependency(dependency)
+    assert package.develop == value
+
+
+def test_search_for_vcs_setup_egg_info(provider: Provider):
+    dependency = VCSDependency("demo", "git", "https://github.com/demo/demo.git")
+
+    package = provider.search_for_direct_origin_dependency(dependency)
+
+    assert package.name == "demo"
+    assert package.version.text == "0.1.2"
+
+    required = [r for r in package.requires if not r.is_optional()]
+    optional = [r for r in package.requires if r.is_optional()]
+    assert required == [get_dependency("pendulum", ">=1.4.4")]
+    assert optional == [get_dependency("tomlkit"), get_dependency("cleo")]
+    assert package.extras == {
+        "foo": [get_dependency("cleo")],
+        "bar": [get_dependency("tomlkit")],
+    }
+
+
+def test_search_for_vcs_setup_egg_info_with_extras(provider: Provider):
+    dependency = VCSDependency(
+        "demo", "git", "https://github.com/demo/demo.git", extras=["foo"]
+    )
+
+    package = provider.search_for_direct_origin_dependency(dependency)
+
+    assert package.name == "demo"
+    assert package.version.text == "0.1.2"
+
+    required = [r for r in package.requires if not r.is_optional()]
+    optional = [r for r in package.requires if r.is_optional()]
+    assert required == [get_dependency("pendulum", ">=1.4.4")]
+    assert optional == [get_dependency("tomlkit"), get_dependency("cleo")]
+    assert package.extras == {
+        "foo": [get_dependency("cleo")],
+        "bar": [get_dependency("tomlkit")],
+    }
+
+
+def test_search_for_vcs_read_setup(provider: Provider, mocker: MockerFixture):
+    mocker.patch("poetry.utils.env.EnvManager.get", return_value=MockEnv())
+
+    dependency = VCSDependency("demo", "git", "https://github.com/demo/demo.git")
+
+    package = provider.search_for_direct_origin_dependency(dependency)
+
+    assert package.name == "demo"
+    assert package.version.text == "0.1.2"
+
+    required = [r for r in package.requires if not r.is_optional()]
+    optional = [r for r in package.requires if r.is_optional()]
+    assert required == [get_dependency("pendulum", ">=1.4.4")]
+    assert optional == [get_dependency("tomlkit"), get_dependency("cleo")]
+    assert package.extras == {
+        "foo": [get_dependency("cleo")],
+        "bar": [get_dependency("tomlkit")],
+    }
+
+
+def test_search_for_vcs_read_setup_with_extras(
+    provider: Provider, mocker: MockerFixture
+):
+    mocker.patch("poetry.utils.env.EnvManager.get", return_value=MockEnv())
+
+    dependency = VCSDependency(
+        "demo", "git", "https://github.com/demo/demo.git", extras=["foo"]
+    )
+
+    package = provider.search_for_direct_origin_dependency(dependency)
+
+    assert package.name == "demo"
+    assert package.version.text == "0.1.2"
+
+    required = [r for r in package.requires if not r.is_optional()]
+    optional = [r for r in package.requires if r.is_optional()]
+    assert required == [get_dependency("pendulum", ">=1.4.4")]
+    assert optional == [get_dependency("tomlkit"), get_dependency("cleo")]
+
+
+def test_search_for_vcs_read_setup_raises_error_if_no_version(
+    provider: Provider, mocker: MockerFixture
+):
+    mocker.patch(
+        "poetry.inspection.info.get_pep517_metadata",
         return_value=PackageInfo(name="demo", version=None),
     )
 
     dependency = VCSDependency("demo", "git", "https://github.com/demo/no-version.git")
 
     with pytest.raises(RuntimeError):
-        provider.search_for_vcs(dependency)
+        provider.search_for_direct_origin_dependency(dependency)
 
 
 @pytest.mark.parametrize("directory", ["demo", "non-canonical-name"])
-def test_search_for_directory_setup_egg_info(provider, directory):
+def test_search_for_directory_setup_egg_info(provider: Provider, directory: str):
     dependency = DirectoryDependency(
         "demo",
         Path(__file__).parent.parent
@@ -149,7 +276,7 @@ def test_search_for_directory_setup_egg_info(provider, directory):
         / directory,
     )
 
-    package = provider.search_for_directory(dependency)[0]
+    package = provider.search_for_direct_origin_dependency(dependency)
 
     assert package.name == "demo"
     assert package.version.text == "0.1.2"
@@ -164,7 +291,7 @@ def test_search_for_directory_setup_egg_info(provider, directory):
     }
 
 
-def test_search_for_directory_setup_egg_info_with_extras(provider):
+def test_search_for_directory_setup_egg_info_with_extras(provider: Provider):
     dependency = DirectoryDependency(
         "demo",
         Path(__file__).parent.parent
@@ -176,7 +303,7 @@ def test_search_for_directory_setup_egg_info_with_extras(provider):
         extras=["foo"],
     )
 
-    package = provider.search_for_directory(dependency)[0]
+    package = provider.search_for_direct_origin_dependency(dependency)
 
     assert package.name == "demo"
     assert package.version.text == "0.1.2"
@@ -192,7 +319,7 @@ def test_search_for_directory_setup_egg_info_with_extras(provider):
 
 
 @pytest.mark.parametrize("directory", ["demo", "non-canonical-name"])
-def test_search_for_directory_setup_with_base(provider, directory):
+def test_search_for_directory_setup_with_base(provider: Provider, directory: str):
     dependency = DirectoryDependency(
         "demo",
         Path(__file__).parent.parent
@@ -209,7 +336,7 @@ def test_search_for_directory_setup_with_base(provider, directory):
         / directory,
     )
 
-    package = provider.search_for_directory(dependency)[0]
+    package = provider.search_for_direct_origin_dependency(dependency)
 
     assert package.name == "demo"
     assert package.version.text == "0.1.2"
@@ -232,8 +359,9 @@ def test_search_for_directory_setup_with_base(provider, directory):
     )
 
 
-@pytest.mark.skipif(not PY35, reason="AST parsing does not work for Python <3.4")
-def test_search_for_directory_setup_read_setup(provider, mocker):
+def test_search_for_directory_setup_read_setup(
+    provider: Provider, mocker: MockerFixture
+):
     mocker.patch("poetry.utils.env.EnvManager.get", return_value=MockEnv())
 
     dependency = DirectoryDependency(
@@ -246,7 +374,7 @@ def test_search_for_directory_setup_read_setup(provider, mocker):
         / "demo",
     )
 
-    package = provider.search_for_directory(dependency)[0]
+    package = provider.search_for_direct_origin_dependency(dependency)
 
     assert package.name == "demo"
     assert package.version.text == "0.1.2"
@@ -261,8 +389,9 @@ def test_search_for_directory_setup_read_setup(provider, mocker):
     }
 
 
-@pytest.mark.skipif(not PY35, reason="AST parsing does not work for Python <3.4")
-def test_search_for_directory_setup_read_setup_with_extras(provider, mocker):
+def test_search_for_directory_setup_read_setup_with_extras(
+    provider: Provider, mocker: MockerFixture
+):
     mocker.patch("poetry.utils.env.EnvManager.get", return_value=MockEnv())
 
     dependency = DirectoryDependency(
@@ -276,7 +405,7 @@ def test_search_for_directory_setup_read_setup_with_extras(provider, mocker):
         extras=["foo"],
     )
 
-    package = provider.search_for_directory(dependency)[0]
+    package = provider.search_for_direct_origin_dependency(dependency)
 
     assert package.name == "demo"
     assert package.version.text == "0.1.2"
@@ -291,8 +420,7 @@ def test_search_for_directory_setup_read_setup_with_extras(provider, mocker):
     }
 
 
-@pytest.mark.skipif(not PY35, reason="AST parsing does not work for Python <3.4")
-def test_search_for_directory_setup_read_setup_with_no_dependencies(provider):
+def test_search_for_directory_setup_read_setup_with_no_dependencies(provider: Provider):
     dependency = DirectoryDependency(
         "demo",
         Path(__file__).parent.parent
@@ -303,7 +431,7 @@ def test_search_for_directory_setup_read_setup_with_no_dependencies(provider):
         / "no-dependencies",
     )
 
-    package = provider.search_for_directory(dependency)[0]
+    package = provider.search_for_direct_origin_dependency(dependency)
 
     assert package.name == "demo"
     assert package.version.text == "0.1.2"
@@ -311,13 +439,13 @@ def test_search_for_directory_setup_read_setup_with_no_dependencies(provider):
     assert package.extras == {}
 
 
-def test_search_for_directory_poetry(provider):
+def test_search_for_directory_poetry(provider: Provider):
     dependency = DirectoryDependency(
         "project-with-extras",
         Path(__file__).parent.parent / "fixtures" / "project_with_extras",
     )
 
-    package = provider.search_for_directory(dependency)[0]
+    package = provider.search_for_direct_origin_dependency(dependency)
 
     assert package.name == "project-with-extras"
     assert package.version.text == "1.2.3"
@@ -339,14 +467,14 @@ def test_search_for_directory_poetry(provider):
     }
 
 
-def test_search_for_directory_poetry_with_extras(provider):
+def test_search_for_directory_poetry_with_extras(provider: Provider):
     dependency = DirectoryDependency(
         "project-with-extras",
         Path(__file__).parent.parent / "fixtures" / "project_with_extras",
         extras=["extras_a"],
     )
 
-    package = provider.search_for_directory(dependency)[0]
+    package = provider.search_for_direct_origin_dependency(dependency)
 
     assert package.name == "project-with-extras"
     assert package.version.text == "1.2.3"
@@ -368,7 +496,7 @@ def test_search_for_directory_poetry_with_extras(provider):
     }
 
 
-def test_search_for_file_sdist(provider):
+def test_search_for_file_sdist(provider: Provider):
     dependency = FileDependency(
         "demo",
         Path(__file__).parent.parent
@@ -377,7 +505,7 @@ def test_search_for_file_sdist(provider):
         / "demo-0.1.0.tar.gz",
     )
 
-    package = provider.search_for_file(dependency)[0]
+    package = provider.search_for_direct_origin_dependency(dependency)
 
     assert package.name == "demo"
     assert package.version.text == "0.1.0"
@@ -399,7 +527,7 @@ def test_search_for_file_sdist(provider):
     }
 
 
-def test_search_for_file_sdist_with_extras(provider):
+def test_search_for_file_sdist_with_extras(provider: Provider):
     dependency = FileDependency(
         "demo",
         Path(__file__).parent.parent
@@ -409,7 +537,7 @@ def test_search_for_file_sdist_with_extras(provider):
         extras=["foo"],
     )
 
-    package = provider.search_for_file(dependency)[0]
+    package = provider.search_for_direct_origin_dependency(dependency)
 
     assert package.name == "demo"
     assert package.version.text == "0.1.0"
@@ -431,7 +559,7 @@ def test_search_for_file_sdist_with_extras(provider):
     }
 
 
-def test_search_for_file_wheel(provider):
+def test_search_for_file_wheel(provider: Provider):
     dependency = FileDependency(
         "demo",
         Path(__file__).parent.parent
@@ -440,7 +568,7 @@ def test_search_for_file_wheel(provider):
         / "demo-0.1.0-py2.py3-none-any.whl",
     )
 
-    package = provider.search_for_file(dependency)[0]
+    package = provider.search_for_direct_origin_dependency(dependency)
 
     assert package.name == "demo"
     assert package.version.text == "0.1.0"
@@ -462,7 +590,7 @@ def test_search_for_file_wheel(provider):
     }
 
 
-def test_search_for_file_wheel_with_extras(provider):
+def test_search_for_file_wheel_with_extras(provider: Provider):
     dependency = FileDependency(
         "demo",
         Path(__file__).parent.parent
@@ -472,7 +600,7 @@ def test_search_for_file_wheel_with_extras(provider):
         extras=["foo"],
     )
 
-    package = provider.search_for_file(dependency)[0]
+    package = provider.search_for_direct_origin_dependency(dependency)
 
     assert package.name == "demo"
     assert package.version.text == "0.1.0"
@@ -491,4 +619,72 @@ def test_search_for_file_wheel_with_extras(provider):
     assert package.extras == {
         "foo": [get_dependency("cleo")],
         "bar": [get_dependency("tomlkit")],
+    }
+
+
+def test_complete_package_preserves_source_type(
+    provider: Provider, root: ProjectPackage
+) -> None:
+    fixtures = Path(__file__).parent.parent / "fixtures"
+    project_dir = fixtures.joinpath("with_conditional_path_deps")
+    for folder in ["demo_one", "demo_two"]:
+        path = (project_dir / folder).as_posix()
+        root.add_dependency(Factory.create_dependency("demo", {"path": path}))
+
+    complete_package = provider.complete_package(
+        DependencyPackage(root.to_dependency(), root)
+    )
+
+    requires = complete_package.package.all_requires
+    assert len(requires) == 2
+    assert {requires[0].source_url, requires[1].source_url} == {
+        project_dir.joinpath("demo_one").as_posix(),
+        project_dir.joinpath("demo_two").as_posix(),
+    }
+
+
+def test_complete_package_preserves_source_type_with_subdirectories(
+    provider: Provider, root: ProjectPackage
+) -> None:
+    dependency_one = Factory.create_dependency(
+        "one",
+        {
+            "git": "https://github.com/demo/subdirectories.git",
+            "subdirectory": "one",
+        },
+    )
+    dependency_one_copy = Factory.create_dependency(
+        "one",
+        {
+            "git": "https://github.com/demo/subdirectories.git",
+            "subdirectory": "one-copy",
+        },
+    )
+    dependency_two = Factory.create_dependency(
+        "two",
+        {"git": "https://github.com/demo/subdirectories.git", "subdirectory": "two"},
+    )
+
+    root.add_dependency(
+        Factory.create_dependency(
+            "one",
+            {
+                "git": "https://github.com/demo/subdirectories.git",
+                "subdirectory": "one",
+            },
+        )
+    )
+    root.add_dependency(dependency_one_copy)
+    root.add_dependency(dependency_two)
+
+    complete_package = provider.complete_package(
+        DependencyPackage(root.to_dependency(), root)
+    )
+
+    requires = complete_package.package.all_requires
+    assert len(requires) == 3
+    assert {r.to_pep_508() for r in requires} == {
+        dependency_one.to_pep_508(),
+        dependency_one_copy.to_pep_508(),
+        dependency_two.to_pep_508(),
     }
