@@ -1,7 +1,8 @@
+from __future__ import annotations
+
 from poetry.core.utils.helpers import parse_requires
-from poetry.utils._compat import Path
-from poetry.utils.helpers import get_cert
-from poetry.utils.helpers import get_client_cert
+
+from poetry.utils.helpers import safe_extra
 
 
 def test_parse_requires():
@@ -20,7 +21,8 @@ msgpack-python>=0.5.0.0,<0.6.0.0
 pyparsing>=2.2.0.0,<3.0.0.0
 requests-toolbelt>=0.8.0.0,<0.9.0.0
 
-[:(python_version >= "2.7.0.0" and python_version < "2.8.0.0") or (python_version >= "3.4.0.0" and python_version < "3.5.0.0")]
+[:(python_version >= "2.7.0.0" and python_version < "2.8.0.0")\
+ or (python_version >= "3.4.0.0" and python_version < "3.5.0.0")]
 typing>=3.6.0.0,<4.0.0.0
 
 [:python_version >= "2.7.0.0" and python_version < "2.8.0.0"]
@@ -32,8 +34,9 @@ zipfile36>=0.1.0.0,<0.2.0.0
 
 [dev]
 isort@ git+git://github.com/timothycrosley/isort.git@e63ae06ec7d70b06df9e528357650281a3d3ec22#egg=isort
-"""
+"""  # noqa: E501
     result = parse_requires(requires)
+    # fmt: off
     expected = [
         "jsonschema>=2.6.0.0,<3.0.0.0",
         "lockfile>=0.12.0.0,<0.13.0.0",
@@ -48,24 +51,18 @@ isort@ git+git://github.com/timothycrosley/isort.git@e63ae06ec7d70b06df9e5283576
         "msgpack-python>=0.5.0.0,<0.6.0.0",
         "pyparsing>=2.2.0.0,<3.0.0.0",
         "requests-toolbelt>=0.8.0.0,<0.9.0.0",
-        'typing>=3.6.0.0,<4.0.0.0 ; (python_version >= "2.7.0.0" and python_version < "2.8.0.0") or (python_version >= "3.4.0.0" and python_version < "3.5.0.0")',
-        'virtualenv>=15.2.0.0,<16.0.0.0 ; python_version >= "2.7.0.0" and python_version < "2.8.0.0"',
-        'pathlib2>=2.3.0.0,<3.0.0.0 ; python_version >= "2.7.0.0" and python_version < "2.8.0.0"',
-        'zipfile36>=0.1.0.0,<0.2.0.0 ; python_version >= "3.4.0.0" and python_version < "3.6.0.0"',
-        'isort@ git+git://github.com/timothycrosley/isort.git@e63ae06ec7d70b06df9e528357650281a3d3ec22#egg=isort ; extra == "dev"',
+        'typing>=3.6.0.0,<4.0.0.0 ; (python_version >= "2.7.0.0" and python_version < "2.8.0.0") or (python_version >= "3.4.0.0" and python_version < "3.5.0.0")',  # noqa: E501
+        'virtualenv>=15.2.0.0,<16.0.0.0 ; python_version >= "2.7.0.0" and python_version < "2.8.0.0"',  # noqa: E501
+        'pathlib2>=2.3.0.0,<3.0.0.0 ; python_version >= "2.7.0.0" and python_version < "2.8.0.0"',  # noqa: E501
+        'zipfile36>=0.1.0.0,<0.2.0.0 ; python_version >= "3.4.0.0" and python_version < "3.6.0.0"',  # noqa: E501
+        'isort@ git+git://github.com/timothycrosley/isort.git@e63ae06ec7d70b06df9e528357650281a3d3ec22#egg=isort ; extra == "dev"',  # noqa: E501
     ]
+    # fmt: on
     assert result == expected
 
 
-def test_get_cert(config):
-    ca_cert = "path/to/ca.pem"
-    config.merge({"certificates": {"foo": {"cert": ca_cert}}})
-
-    assert get_cert(config, "foo") == Path(ca_cert)
-
-
-def test_get_client_cert(config):
-    client_cert = "path/to/client.pem"
-    config.merge({"certificates": {"foo": {"client-cert": client_cert}}})
-
-    assert get_client_cert(config, "foo") == Path(client_cert)
+def test_safe_extra():
+    extra = "pandas.CSVDataSet"
+    result = safe_extra(extra)
+    expected = "pandas.csvdataset"
+    assert result == expected
