@@ -1,21 +1,31 @@
+from __future__ import annotations
+
 import os
+
+from typing import TYPE_CHECKING
 
 import pytest
 
-from poetry.utils._compat import PY35
+from poetry.core.version.exceptions import InvalidVersion
+
 from poetry.utils.setup_reader import SetupReader
 
 
+if TYPE_CHECKING:
+    from collections.abc import Callable
+
+
 @pytest.fixture()
-def setup():
-    def _setup(name):
+def setup() -> Callable[[str], str]:
+    def _setup(name: str) -> str:
         return os.path.join(os.path.dirname(__file__), "fixtures", "setups", name)
 
     return _setup
 
 
-@pytest.mark.skipif(not PY35, reason="AST parsing does not work for Python <3.4")
-def test_setup_reader_read_first_level_setup_call_with_direct_types(setup):
+def test_setup_reader_read_first_level_setup_call_with_direct_types(
+    setup: Callable[[str], str]
+):
     result = SetupReader.read_from_directory(setup("flask"))
 
     expected_name = "Flask"
@@ -40,15 +50,16 @@ def test_setup_reader_read_first_level_setup_call_with_direct_types(setup):
     }
     expected_python_requires = ">=2.7,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*"
 
-    assert expected_name == result["name"]
-    assert expected_version == result["version"]
-    assert expected_install_requires == result["install_requires"]
-    assert expected_extras_require == result["extras_require"]
-    assert expected_python_requires == result["python_requires"]
+    assert result["name"] == expected_name
+    assert result["version"] == expected_version
+    assert result["install_requires"] == expected_install_requires
+    assert result["extras_require"] == expected_extras_require
+    assert result["python_requires"] == expected_python_requires
 
 
-@pytest.mark.skipif(not PY35, reason="AST parsing does not work for Python <3.4")
-def test_setup_reader_read_first_level_setup_call_with_variables(setup):
+def test_setup_reader_read_first_level_setup_call_with_variables(
+    setup: Callable[[str], str]
+):
     result = SetupReader.read_from_directory(setup("requests"))
 
     expected_name = None
@@ -66,15 +77,16 @@ def test_setup_reader_read_first_level_setup_call_with_variables(setup):
     }
     expected_python_requires = ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*"
 
-    assert expected_name == result["name"]
-    assert expected_version == result["version"]
-    assert expected_install_requires == result["install_requires"]
-    assert expected_extras_require == result["extras_require"]
-    assert expected_python_requires == result["python_requires"]
+    assert result["name"] == expected_name
+    assert result["version"] == expected_version
+    assert result["install_requires"] == expected_install_requires
+    assert result["extras_require"] == expected_extras_require
+    assert result["python_requires"] == expected_python_requires
 
 
-@pytest.mark.skipif(not PY35, reason="AST parsing does not work for Python <3.4")
-def test_setup_reader_read_sub_level_setup_call_with_direct_types(setup):
+def test_setup_reader_read_sub_level_setup_call_with_direct_types(
+    setup: Callable[[str], str]
+):
     result = SetupReader.read_from_directory(setup("sqlalchemy"))
 
     expected_name = "SQLAlchemy"
@@ -91,14 +103,14 @@ def test_setup_reader_read_sub_level_setup_call_with_direct_types(setup):
         "mssql_pymssql": ["pymssql"],
     }
 
-    assert expected_name == result["name"]
-    assert expected_version == result["version"]
-    assert expected_install_requires == result["install_requires"]
-    assert expected_extras_require == result["extras_require"]
+    assert result["name"] == expected_name
+    assert result["version"] == expected_version
+    assert result["install_requires"] == expected_install_requires
+    assert result["extras_require"] == expected_extras_require
     assert result["python_requires"] is None
 
 
-def test_setup_reader_read_setup_cfg(setup):
+def test_setup_reader_read_setup_cfg(setup: Callable[[str], str]):
     result = SetupReader.read_from_directory(setup("with-setup-cfg"))
 
     expected_name = "with-setup-cfg"
@@ -110,15 +122,19 @@ def test_setup_reader_read_setup_cfg(setup):
     }
     expected_python_requires = ">=2.6,!=3.0,!=3.1,!=3.2,!=3.3"
 
-    assert expected_name == result["name"]
-    assert expected_version == result["version"]
-    assert expected_install_requires == result["install_requires"]
-    assert expected_extras_require == result["extras_require"]
-    assert expected_python_requires == result["python_requires"]
+    assert result["name"] == expected_name
+    assert result["version"] == expected_version
+    assert result["install_requires"] == expected_install_requires
+    assert result["extras_require"] == expected_extras_require
+    assert result["python_requires"] == expected_python_requires
 
 
-@pytest.mark.skipif(not PY35, reason="AST parsing does not work for Python <3.4")
-def test_setup_reader_read_setup_kwargs(setup):
+def test_setup_reader_read_setup_cfg_with_attr(setup: Callable[[str], str]):
+    with pytest.raises(InvalidVersion):
+        SetupReader.read_from_directory(setup("with-setup-cfg-attr"))
+
+
+def test_setup_reader_read_setup_kwargs(setup: Callable[[str], str]):
     result = SetupReader.read_from_directory(setup("pendulum"))
 
     expected_name = "pendulum"
@@ -127,15 +143,14 @@ def test_setup_reader_read_setup_kwargs(setup):
     expected_extras_require = {':python_version < "3.5"': ["typing>=3.6,<4.0"]}
     expected_python_requires = ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*"
 
-    assert expected_name == result["name"]
-    assert expected_version == result["version"]
-    assert expected_install_requires == result["install_requires"]
-    assert expected_extras_require == result["extras_require"]
-    assert expected_python_requires == result["python_requires"]
+    assert result["name"] == expected_name
+    assert result["version"] == expected_version
+    assert result["install_requires"] == expected_install_requires
+    assert result["extras_require"] == expected_extras_require
+    assert result["python_requires"] == expected_python_requires
 
 
-@pytest.mark.skipif(not PY35, reason="AST parsing does not work for Python <3.4")
-def test_setup_reader_read_setup_call_in_main(setup):
+def test_setup_reader_read_setup_call_in_main(setup: Callable[[str], str]):
     result = SetupReader.read_from_directory(setup("pyyaml"))
 
     expected_name = "PyYAML"
@@ -144,15 +159,14 @@ def test_setup_reader_read_setup_call_in_main(setup):
     expected_extras_require = {}
     expected_python_requires = None
 
-    assert expected_name == result["name"]
-    assert expected_version == result["version"]
-    assert expected_install_requires == result["install_requires"]
-    assert expected_extras_require == result["extras_require"]
-    assert expected_python_requires == result["python_requires"]
+    assert result["name"] == expected_name
+    assert result["version"] == expected_version
+    assert result["install_requires"] == expected_install_requires
+    assert result["extras_require"] == expected_extras_require
+    assert result["python_requires"] == expected_python_requires
 
 
-@pytest.mark.skipif(not PY35, reason="AST parsing does not work for Python <3.4")
-def test_setup_reader_read_extras_require_with_variables(setup):
+def test_setup_reader_read_extras_require_with_variables(setup: Callable[[str], str]):
     result = SetupReader.read_from_directory(setup("extras_require_with_vars"))
 
     expected_name = "extras_require_with_vars"
@@ -161,19 +175,18 @@ def test_setup_reader_read_extras_require_with_variables(setup):
     expected_extras_require = {"test": ["pytest"]}
     expected_python_requires = None
 
-    assert expected_name == result["name"]
-    assert expected_version == result["version"]
-    assert expected_install_requires == result["install_requires"]
-    assert expected_extras_require == result["extras_require"]
-    assert expected_python_requires == result["python_requires"]
+    assert result["name"] == expected_name
+    assert result["version"] == expected_version
+    assert result["install_requires"] == expected_install_requires
+    assert result["extras_require"] == expected_extras_require
+    assert result["python_requires"] == expected_python_requires
 
 
-@pytest.mark.skipif(not PY35, reason="AST parsing does not work for Python <3.4")
-def test_setup_reader_setuptools(setup):
+def test_setup_reader_setuptools(setup: Callable[[str], str]):
     result = SetupReader.read_from_directory(setup("setuptools_setup"))
 
     expected_name = "my_package"
     expected_version = "0.1.2"
 
-    assert expected_name == result["name"]
-    assert expected_version == result["version"]
+    assert result["name"] == expected_name
+    assert result["version"] == expected_version
